Refactor maps database definitions
