+++ src/db/maps/maps-db.constants.ts
@@ -1,0 +1,3 @@
+import type { SceneVector2 } from "@core/logic/provided/services/scene-object-manager/scene-object-manager.types";
+
+export const FOUNDATIONS_CENTER: SceneVector2 = { x: 500, y: 500 };
