+++ src/db/maps/maps-db.ts
@@ -1,3962 +1,70 @@
-import { BrickType, getBrickConfig } from "../bricks-db";
-import {
-  SceneSize,
-  SceneVector2,
-} from "@core/logic/provided/services/scene-object-manager/scene-object-manager.types";
-import { PlayerUnitType } from "../player-units-db";
-import type { EnemyType } from "../enemies-db";
-import type { EnemySpawnData } from "../../logic/modules/active-map/enemies/enemies.types";
-import type { UnlockCondition } from "@shared/types/unlocks";
-import type { SkillId } from "../skills-db";
-import type { AchievementId } from "../achievements-db";
-import type { MapEffectId } from "../map-effects-db";
-import {
-  BrickShapeBlueprint,
-  buildBricksFromBlueprints,
-  bezierCurveWithBricks,
-  bezierPolygonWithBricks,
-  circleWithBricks,
-  connectorWithBricks,
-  polygonWithBricks,
-  squareWithBricks,
-  templateWithBricks,
-} from "../../logic/services/brick-layout/BrickLayoutService";
-import { transformBezierOutline } from "../../logic/services/brick-layout/brick-layout.helpers";
-import { brickTreeDeterministic } from "./helpers/bush-helper";
+import type { MapConfig, MapId, MapListEntry } from "./maps-db.types";
+import adit from "./map-definitions/adit";
+import ancientPyramids from "./map-definitions/ancientPyramids";
+import bezierGrove from "./map-definitions/bezierGrove";
+import coil from "./map-definitions/coil";
+import deadOak from "./map-definitions/deadOak";
+import deadlyTunnels from "./map-definitions/deadlyTunnels";
+import deathfulGuns from "./map-definitions/deathfulGuns";
+import encagedBeast from "./map-definitions/encagedBeast";
+import foundations from "./map-definitions/foundations";
+import frozenForest from "./map-definitions/frozenForest";
+import initial from "./map-definitions/initial";
+import megaBrick from "./map-definitions/megaBrick";
+import mine from "./map-definitions/mine";
+import oldForge from "./map-definitions/oldForge";
+import silverRing from "./map-definitions/silverRing";
+import sphinx from "./map-definitions/sphinx";
+import spruce from "./map-definitions/spruce";
+import stoneCottage from "./map-definitions/stoneCottage";
+import thicket from "./map-definitions/thicket";
+import trainingGrounds from "./map-definitions/trainingGrounds";
+import turretRings from "./map-definitions/turretRings";
+import tutorialZone from "./map-definitions/tutorialZone";
+import uraniumFields from "./map-definitions/uranium_fields";
+import volcano from "./map-definitions/volcano";
+import wire from "./map-definitions/wire";
 
-export type MapId =
-  | "tutorialZone"
-  | "trainingGrounds"
-  | "foundations"
-  | "initial"
-  | "turretRings"
-  | "thicket"
-  | "oldForge"
-  | "spruce"
-  | "deadOak"
-  | "sphinx"
-  | "stoneCottage"
-  | "bezierGrove"
-  | "wire"
-  | "coil"
-  | "mine"
-  | "adit"
-  | "silverRing"
-  | "frozenForest"
-  | "volcano"
-  | "megaBrick"
-  | "ancientPyramids"
-  | "deathfulGuns"
-  | "deadlyTunnels"
-  | "encagedBeast"
-  | "uranium_fields";
-
-export interface MapBrickGeneratorOptions {
-  readonly mapLevel: number;
-}
-
-export type MapBrickGenerator = (
-  options: MapBrickGeneratorOptions,
-) => readonly BrickShapeBlueprint[];
-
-export interface MapEnemyGeneratorOptions {
-  readonly mapLevel: number;
-}
-
-export type MapEnemyGenerator = (
-  options: MapEnemyGeneratorOptions,
-) => readonly EnemySpawnData[];
-
-export interface MapNodePosition {
-  readonly x: number;
-  readonly y: number;
-}
-
-export interface MapEnemySpawnTypeConfig {
-  readonly type: EnemyType;
-  readonly weight: number; // Вага для випадкового вибору (1.0 = базовий, 2.0 = вдвічі частіше)
-  readonly minLevel?: number; // Мінімальний рівень карти для появи
-  readonly maxLevel?: number; // Максимальний рівень карти
-}
-
-export interface MapEnemySpawnPointConfig {
-  readonly position: SceneVector2;
-  readonly spawnRate: number; // Ворогів на секунду (або інтервал між спавнами)
-  readonly enemyTypes: readonly MapEnemySpawnTypeConfig[];
-  readonly maxConcurrent?: number; // Максимальна кількість одночасно активних ворогів
-  readonly enabled?: boolean; // Можна вимкнути для певних рівнів
-  readonly levelOffset?: number; // Зміщення рівня ворогів відносно рівня карти (за замовчуванням 0)
-}
-
-export interface MapConfig {
-  readonly name: string;
-  readonly size: SceneSize;
-  readonly bricks: MapBrickGenerator;
-  readonly playerUnits?: readonly MapPlayerUnitConfig[];
-  readonly spawnPoints?: readonly SceneVector2[];
-  readonly enemySpawnPoints?: readonly MapEnemySpawnPointConfig[];
-  readonly enemies?: MapEnemyGenerator; // Статичні вороги (турелі), що генеруються один раз при ініціалізації
-  readonly mapEffects?: readonly MapEffectId[];
-  readonly unlockedBy?: readonly UnlockCondition<MapId, SkillId>[];
-  readonly icon?: string;
-  readonly nodePosition: MapNodePosition;
-  readonly mapsRequired?: Partial<Record<MapId, number>>;
-  readonly maxLevel: number;
-  readonly resourceMultiplier?: number; // Множник ресурсів для цієї мапи (застосовується до brick_rewards)
-  readonly achievementId?: AchievementId;
-  readonly lockedForDemo?: boolean;
-}
-
-export interface MapListEntry {
-  readonly id: MapId;
-  readonly name: string;
-  readonly size: SceneSize;
-  readonly icon?: string;
-}
-
-export interface MapPlayerUnitConfig {
-  readonly type: PlayerUnitType;
-  readonly position: SceneVector2;
-}
-
-const FOUNDATIONS_CENTER: SceneVector2 = { x: 500, y: 500 };
+export type {
+  MapBrickGenerator,
+  MapBrickGeneratorOptions,
+  MapConfig,
+  MapEnemyGenerator,
+  MapEnemyGeneratorOptions,
+  MapEnemySpawnPointConfig,
+  MapEnemySpawnTypeConfig,
+  MapId,
+  MapListEntry,
+  MapNodePosition,
+  MapPlayerUnitConfig,
+} from "./maps-db.types";
 
 const MAPS_DB: Record<MapId, MapConfig> = {
-  tutorialZone: (() => {
-    const center: SceneVector2 = { x: 500, y: 600 };
-    const size: SceneSize = { width: 1000, height: 1000 };
-    const spawnPoint: SceneVector2 = { x: center.x, y: center.y - 500 };
-
-    // Простий шаблон цифри "1"
-    const numberOneTemplate: readonly string[] = [
-      " ####      ##      #####",
-      " #        #  #       #",
-      " ####     ####       #",
-      " #       #    #      #",
-      " ####   #      #     #",
-    ];
-
-    return {
-      name: "Weird Bricks",
-      size,
-      spawnPoints: [spawnPoint],
-      icon: "eat.png",
-      bricks: ({ mapLevel }) => {
-        const baseLevel = Math.max(0, Math.floor(mapLevel));
-
-        const numberOne = templateWithBricks(
-          "smallTrainingBrick",
-          {
-            center,
-            template: numberOneTemplate,
-            horizontalGap: 1,
-            verticalGap: 1,
-          },
-          { level: baseLevel },
-        );
-
-        return [numberOne];
-      },
-      playerUnits: [
-        {
-          type: "bluePentagon",
-          position: { ...spawnPoint },
-        },
-      ],
-      nodePosition: { x: -1, y: -1 },
-      maxLevel: 1,
-    } satisfies MapConfig;
-  })(),
-  trainingGrounds: (() => {
-    const center: SceneVector2 = { x: 500, y: 600 };
-    const size: SceneSize = { width: 1000, height: 1000 };
-    const spawnPoint: SceneVector2 = { x: center.x, y: center.y - 500 };
-
-    // Голова смайлика
-    const headRadius = 280;
-    const headThickness = 20;
-
-    // Очі
-    const eyeRadius = 35;
-    const eyeOffsetY = -60;
-    const eyeOffsetX = 80;
-
-    // Рот (дуга) - створюємо через сегменти кіл
-    const mouthRadius = 120;
-    const mouthThickness = headThickness;
-    const mouthCenterY = center.y + 50;
-    const mouthStartAngle = Math.PI * 0.25; // ~45 градусів
-    const mouthEndAngle = Math.PI * 0.75; // ~135 градусів
-    const mouthSegments = 8; // кількість сегментів для рота
-
-    return {
-      name: "Optimistic Smile",
-      size,
-      spawnPoints: [spawnPoint],
-      icon: "training.png",
-      bricks: ({ mapLevel }) => {
-        const baseLevel = Math.max(0, Math.floor(mapLevel));
-
-        // Голова (зовнішнє коло)
-        const headOuter = circleWithBricks(
-          "smallTrainingBrick",
-          {
-            center,
-            innerRadius: headRadius - headThickness,
-            outerRadius: headRadius,
-          },
-          { level: baseLevel },
-        );
-
-        // Ліве око
-        const leftEye = circleWithBricks(
-          "smallTrainingBrick",
-          {
-            center: { x: center.x - eyeOffsetX, y: center.y + eyeOffsetY },
-            innerRadius: 0,
-            outerRadius: eyeRadius,
-          },
-          { level: baseLevel },
-        );
-
-        // Праве око
-        const rightEye = circleWithBricks(
-          "smallTrainingBrick",
-          {
-            center: { x: center.x + eyeOffsetX, y: center.y + eyeOffsetY },
-            innerRadius: 0,
-            outerRadius: eyeRadius,
-          },
-          { level: baseLevel },
-        );
-
-        // Рот (дуга) - створюємо через сегменти кіл
-        const mouthSegmentsArray = Array.from(
-          { length: mouthSegments },
-          (_, i) => {
-            const t = i / (mouthSegments - 1);
-            const angle =
-              mouthStartAngle + (mouthEndAngle - mouthStartAngle) * t;
-            const segmentCenter: SceneVector2 = {
-              x: center.x + Math.cos(angle) * mouthRadius,
-              y: mouthCenterY + Math.sin(angle) * mouthRadius,
-            };
-            return circleWithBricks(
-              "smallTrainingBrick",
-              {
-                center: segmentCenter,
-                innerRadius: 0,
-                outerRadius: mouthThickness,
-              },
-              { level: baseLevel },
-            );
-          },
-        );
-
-        return [headOuter, leftEye, rightEye, ...mouthSegmentsArray];
-      },
-      playerUnits: [
-        {
-          type: "bluePentagon",
-          position: { ...spawnPoint },
-        },
-      ],
-      nodePosition: { x: 0, y: 0 },
-      mapsRequired: { tutorialZone: 1 },
-      maxLevel: 1,
-    } satisfies MapConfig;
-  })(),
-  foundations: (() => {
-    const center = FOUNDATIONS_CENTER;
-    const size: SceneSize = { width: 1000, height: 1000 };
-    const spawnPoint: SceneVector2 = { x: center.x, y: center.y - 30 };
-    const sides = 5;
-    const outerRadius = 360;
-    const layerThicknessTraining =
-      getBrickConfig("smallTrainingBrick").size.width;
-    const layerThicknessGray = getBrickConfig("smallSquareGray").size.width;
-    const innerRadius = Math.max(
-      outerRadius - layerThicknessTraining - layerThicknessGray,
-      0,
-    );
-    const middleRadius = Math.max(outerRadius - layerThicknessGray, 0);
-
-    const createPolygon = (radius: number): SceneVector2[] =>
-      Array.from({ length: sides }, (_, index) => {
-        const angle = (index / sides) * Math.PI * 2 - Math.PI / 2;
-        return {
-          x: center.x + Math.cos(angle) * radius,
-          y: center.y + Math.sin(angle) * radius,
-        };
-      });
-
-    const outerVertices = createPolygon(outerRadius);
-    const innerVertices = createPolygon(innerRadius);
-    const middleVertices = createPolygon(middleRadius);
-    const expandedVertices = createPolygon(
-      outerRadius + getBrickConfig("smallSquareGray").size.width * 1.2,
-    );
-
-    return {
-      name: "Cracked Pentagon",
-      size,
-      icon: "pentagon.png",
-      spawnPoints: [spawnPoint],
-      unlockedBy: [
-        {
-          type: "map",
-          id: "trainingGrounds",
-          level: 1,
-        },
-      ],
-      nodePosition: { x: 1, y: 1 },
-      mapsRequired: { trainingGrounds: 1 },
-      maxLevel: 1,
-      bricks: ({ mapLevel }) => [
-        polygonWithBricks(
-          "smallTrainingBrick",
-          {
-            vertices: middleVertices,
-            holes: [innerVertices],
-            offsetX: center.x,
-            offsetY: center.y,
-          },
-          { level: mapLevel },
-        ),
-        polygonWithBricks(
-          "smallSquareGray",
-          {
-            vertices: outerVertices,
-            holes: [middleVertices],
-            offsetX: center.x,
-            offsetY: center.y,
-          },
-          { level: mapLevel },
-        ),
-        ...expandedVertices.map((vertex) =>
-          circleWithBricks(
-            "smallSquareGray",
-            {
-              center: vertex,
-              innerRadius: 0,
-              outerRadius: getBrickConfig("smallSquareGray").size.width * 2.3,
-            },
-            { level: mapLevel },
-          ),
-        ),
-      ],
-      playerUnits: [
-        {
-          type: "bluePentagon",
-          position: { ...spawnPoint },
-        },
-      ],
-    } satisfies MapConfig;
-  })(),
-  megaBrick: (() => {
-    const center = FOUNDATIONS_CENTER;
-    const size: SceneSize = { width: 1000, height: 1000 };
-    const spawnPoint: SceneVector2 = { x: 100, y: center.y - 30 };
-
-    // Один цегла по центру
-    const singleBrickTemplate: readonly string[] = [" # "];
-
-    return {
-      name: "Mega Brick",
-      size,
-      icon: "mega_brick.png",
-      spawnPoints: [spawnPoint],
-      unlockedBy: [
-        {
-          type: "map",
-          id: "trainingGrounds",
-          level: 1,
-        },
-      ],
-      nodePosition: { x: -1, y: 1 },
-      maxLevel: 10,
-      achievementId: "megaBrick",
-      bricks: ({ mapLevel }) => [
-        templateWithBricks(
-          "megaBrick",
-          {
-            center,
-            template: singleBrickTemplate,
-          },
-          { level: mapLevel },
-        ),
-      ],
-      playerUnits: [
-        {
-          type: "bluePentagon",
-          position: { ...spawnPoint },
-        },
-      ],
-      mapsRequired: { trainingGrounds: 1 },
-    } satisfies MapConfig;
-  })(),
-  ancientPyramids: (() => {
-    const center = FOUNDATIONS_CENTER;
-    const size: SceneSize = { width: 1200, height: 1200 };
-    const spawnPoint: SceneVector2 = { x: 150, y: 150 };
-    const pyramidTemplate: readonly string[] = [
-      "           #           ",
-      "          ###          ",
-      "         #####         ",
-      "        #######        ",
-      "       #########       ",
-      "      ###########      ",
-      "     #############     ",
-      "    ###############    ",
-      "   #################   ",
-      "  ###################  ",
-      " ##################### ",
-      "#######################",
-    ];
-
-    return {
-      name: "Ancient Piramids",
-      size,
-      icon: "pyramids.png",
-      spawnPoints: [spawnPoint],
-      unlockedBy: [
-        {
-          type: "map",
-          id: "megaBrick",
-          level: 1,
-        },
-      ],
-      nodePosition: { x: -2, y: 1 },
-      maxLevel: 10,
-      achievementId: "ancientPyramids",
-      bricks: ({ mapLevel }) => {
-        const pyramidLevel = Math.max(1, Math.floor(mapLevel)) + 1;
-        return [
-          templateWithBricks(
-            "smallSquareYellow",
-            {
-              center,
-              template: pyramidTemplate,
-              horizontalGap: 1,
-              verticalGap: 1,
-            },
-            { level: pyramidLevel },
-          ),
-        ];
-      },
-      enemies: ({ mapLevel }) => {
-        const level = Math.max(1, Math.floor(mapLevel));
-        const turretCount = 4;
-        const turretRadius = 320;
-        return Array.from({ length: turretCount }, (_, index) => {
-          const angle = (index / turretCount) * Math.PI * 2;
-          const position: SceneVector2 = {
-            x: center.x + Math.cos(angle) * turretRadius,
-            y: center.y + Math.sin(angle) * turretRadius,
-          };
-          return {
-            type: "freezeTurretEnemy",
-            level,
-            position,
-          } satisfies EnemySpawnData;
-        });
-      },
-      playerUnits: [
-        {
-          type: "bluePentagon",
-          position: { ...spawnPoint },
-        },
-      ],
-      mapsRequired: { megaBrick: 1 },
-    } satisfies MapConfig;
-  })(),
-  deathfulGuns: (() => {
-    const size: SceneSize = { width: 1200, height: 1200 };
-    const center: SceneVector2 = { x: 600, y: 600 };
-    const spawnPoint: SceneVector2 = { x: 600, y: 600 };
-
-    const satelliteCount = 8;
-    const satelliteRadius = 80;
-    const orbitRadius = 350 + satelliteRadius;
-
-    const satellites = Array.from({ length: satelliteCount }, (_, index) => {
-      const angle = (index / satelliteCount) * Math.PI * 2;
-      const position: SceneVector2 = {
-        x: center.x + Math.cos(angle) * orbitRadius,
-        y: center.y + Math.sin(angle) * orbitRadius,
-      };
-      return position;
-    });
-
-    return {
-      name: "Deathful Guns",
-      size,
-      icon: "turrets_of_death.png",
-      spawnPoints: [spawnPoint],
-      unlockedBy: [
-        {
-          type: "map",
-          id: "ancientPyramids",
-          level: 1,
-        },
-      ],
-      nodePosition: { x: -3, y: 1 },
-      maxLevel: 10,
-      achievementId: "deathfulGuns",
-      bricks: ({ mapLevel }) => {
-        
-        
-        const satelliteCount = 8;
-        const satelliteRadius = 80;
-        const orbitRadius = 350 + satelliteRadius;
-  
-        const bricks = satellites.map((position) => {
-          return circleWithBricks(
-            "smallIron",
-            {
-              center: position,
-              innerRadius: satelliteRadius * 0.6,
-              outerRadius: satelliteRadius,
-            },
-            { level: mapLevel },
-          );
-        });
-
-        return [
-          ...bricks,
-        ];
-      },
-      enemies: ({ mapLevel }) => {
-        const level = Math.max(1, Math.floor(mapLevel));
-        return satellites.map((position) => {
-          return {
-            type: "bigGun",
-            level,
-            position,
-          } satisfies EnemySpawnData;
-        });
-      },
-      playerUnits: [
-        {
-          type: "bluePentagon",
-          position: { ...spawnPoint },
-        },
-      ],
-      mapsRequired: { ancientPyramids: 1 },
-    } satisfies MapConfig;
-  })(),
-  deadlyTunnels: (() => {
-    const size: SceneSize = { width: 1200, height: 1200 };
-    const center: SceneVector2 = { x: 600, y: 600 };
-    const spawnPoint: SceneVector2 = { x: 600, y: 600 };
-
-    const satelliteCount = 8;
-    const satelliteRadius = 80;
-    const orbitRadius = 350 + satelliteRadius;
-
-    const satellites = Array.from({ length: satelliteCount }, (_, index) => {
-      const angle = (index / satelliteCount) * Math.PI * 2;
-      const position: SceneVector2 = {
-        x: center.x + Math.cos(angle) * orbitRadius,
-        y: center.y + Math.sin(angle) * orbitRadius,
-      };
-      return position;
-    });
-
-    return {
-      name: "Deadly Tunnels",
-      size,
-      icon: "deadly_tunnels.png",
-      spawnPoints: [spawnPoint],
-      unlockedBy: [
-        {
-          type: "map",
-          id: "deathfulGuns",
-          level: 1,
-        },
-      ],
-      nodePosition: { x: -4, y: 0 },
-      maxLevel: 10,
-      bricks: ({ mapLevel }) => {
-        const stoneLevel = mapLevel + 3;
-
-        // Створюємо квадрати з міді
-        const copperSquares = satellites.map((position) => {
-          const squareSize = satelliteRadius * 2;
-          return squareWithBricks(
-            "smallCopper",
-            {
-              center: position,
-              size: squareSize,
-              innerSize: satelliteRadius * 0.9,
-            },
-            { level: mapLevel },
-          );
-        });
-
-        // Створюємо з'єднання з каменю між супутниками
-        const stoneConnectors = satellites.flatMap((position, index) => {
-          const nextIndex = (index + 1) % satelliteCount;
-          const nextPosition = satellites[nextIndex];
-          if (!nextPosition) {
-            return [];
-          }
-
-          // Обчислюємо точки на зовнішніх краях квадратів для з'єднання
-          const squareSize = satelliteRadius * 2;
-          const halfSize = squareSize / 2;
-
-          // Напрямок від першого квадрата до другого
-          const dx = nextPosition.x - position.x;
-          const dy = nextPosition.y - position.y;
-          const length = Math.hypot(dx, dy);
-          
-          if (length === 0) {
-            return [];
-          }
-
-          // Нормалізуємо вектор напрямку
-          const ux = dx / length;
-          const uy = dy / length;
-
-          // Для квадрата без обертання: знаходимо точку на грані
-          // Використовуємо максимальну координату для визначення грані
-          const absUx = Math.abs(ux);
-          const absUy = Math.abs(uy);
-          const maxAbs = Math.max(absUx, absUy);
-
-          // Масштабуємо для досягнення грані квадрата
-          const scale1 = halfSize / maxAbs;
-          const edgePoint1: SceneVector2 = {
-            x: position.x + ux * scale1,
-            y: position.y + uy * scale1,
-          };
-
-          // Аналогічно для другого квадрата (в зворотному напрямку)
-          const scale2 = halfSize / maxAbs;
-          const edgePoint2: SceneVector2 = {
-            x: nextPosition.x - ux * scale2,
-            y: nextPosition.y - uy * scale2,
-          };
-
-          const connectorWidth = satelliteRadius * 0.8;
-          return connectorWithBricks(
-            "smallSquareGray",
-            {
-              start: edgePoint1,
-              end: edgePoint2,
-              width: connectorWidth,
-            },
-            { level: stoneLevel },
-          );
-        });
-
-        return [...copperSquares, ...stoneConnectors];
-      },
-      enemies: ({ mapLevel }) => {
-        const level = Math.max(1, Math.floor(mapLevel));
-        return satellites.map((position) => {
-          return {
-            type: "volleyTurretEnemy",
-            level,
-            position,
-          } satisfies EnemySpawnData;
-        });
-      },
-      playerUnits: [
-        {
-          type: "bluePentagon",
-          position: { ...spawnPoint },
-        },
-      ],
-      mapsRequired: { deathfulGuns: 1 },
-    } satisfies MapConfig;
-  })(),
-  encagedBeast: (() => {
-    const size: SceneSize = { width: 1200, height: 1200 };
-    const center: SceneVector2 = { x: 600, y: 600 };
-    const spawnPoint: SceneVector2 = { x: 600, y: 100 };
-
-    const satelliteCount = 4;
-    const satelliteRadius = 80;
-    const orbitRadius = 350 + satelliteRadius;
-
-    const satellites = Array.from({ length: satelliteCount }, (_, index) => {
-      const angle = (index / satelliteCount) * Math.PI * 2 + Math.PI / 4;
-      const position: SceneVector2 = {
-        x: center.x + Math.cos(angle) * orbitRadius,
-        y: center.y + Math.sin(angle) * orbitRadius,
-      };
-      return position;
-    });
-
-    return {
-      name: "Encaged Beast",
-      size,
-      icon: "encaged_beast.png",
-      spawnPoints: [spawnPoint],
-      unlockedBy: [
-        {
-          type: "map",
-          id: "deathfulGuns",
-          level: 1,
-        },
-      ],
-      nodePosition: { x: -3, y: 2 },
-      maxLevel: 10,
-      bricks: ({ mapLevel }) => {
-        
-        const copperSquares = satellites.map((position) => {
-          const squareSize = satelliteRadius * 2;
-          return squareWithBricks(
-            "smallCopper",
-            {
-              center: position,
-              size: squareSize,
-              innerSize: satelliteRadius,
-            },
-            { level: mapLevel + 1 },
-          );
-        });
-
-        const centerSquare = squareWithBricks(
-          "smallWood",
-          {
-            center: center,
-            size: (orbitRadius - satelliteRadius) * 1.3,
-            innerSize: (orbitRadius - satelliteRadius) * 0.9,
-          },
-          { level: mapLevel + 1 },
-        );
-
-
-        return [...copperSquares, centerSquare];
-      },
-      enemies: ({ mapLevel }) => {
-        const level = Math.max(1, Math.floor(mapLevel));
-        return [...satellites.map((position) => {
-          return {
-            type: "explosionTurretEnemy",
-            level,
-            position,
-          } satisfies EnemySpawnData;
-        }), {
-          type: "encagedBeastEnemy",
-          level,
-          position: center,
-        } satisfies EnemySpawnData];
-      },
-      playerUnits: [
-        {
-          type: "bluePentagon",
-          position: { ...spawnPoint },
-        },
-      ],
-      mapsRequired: { deathfulGuns: 1 },
-    } satisfies MapConfig;
-  })(),
-  initial: {
-    name: "Initial Grounds",
-    size: { width: 1200, height: 1200 },
-    icon: "initial.png",
-    unlockedBy: [
-      {
-        type: "map",
-        id: "foundations",
-        level: 1,
-      },
-    ],
-    nodePosition: { x: 2, y: 2 },
-    bricks: ({ mapLevel }) => {
-      const baseLevel = Math.max(0, Math.floor(mapLevel));
-      const innerLevel = baseLevel + 1;
-      const center: SceneVector2 = { x: 600, y: 600 };
-      const largeCircle = circleWithBricks(
-        "smallSquareGray",
-        {
-          center,
-          innerRadius: 210,
-          outerRadius: 250,
-        },
-        { level: innerLevel },
-      );
-
-      const largeYellowCircle = circleWithBricks(
-        "smallSquareYellow",
-        {
-          center,
-          innerRadius: 150,
-          outerRadius: 210,
-        },
-        { level: baseLevel },
-      );
-
-      const satelliteCount = 8;
-      const satelliteRadius = 80;
-      const orbitRadius = 350 + satelliteRadius;
-
-      const satellites = Array.from({ length: satelliteCount }, (_, index) => {
-        const angle = (index / satelliteCount) * Math.PI * 2;
-        const position: SceneVector2 = {
-          x: center.x + Math.cos(angle) * orbitRadius,
-          y: center.y + Math.sin(angle) * orbitRadius,
-        };
-        return circleWithBricks(
-          "smallSquareGray",
-          {
-            center: position,
-            innerRadius: satelliteRadius * 0.6,
-            outerRadius: satelliteRadius,
-          },
-          { level: baseLevel + 0.5 },
-        );
-      });
-
-      const satellitesInner = Array.from(
-        { length: satelliteCount },
-        (_, index) => {
-          const angle = (index / satelliteCount) * Math.PI * 2;
-          const position: SceneVector2 = {
-            x: center.x + Math.cos(angle) * orbitRadius,
-            y: center.y + Math.sin(angle) * orbitRadius,
-          };
-          return circleWithBricks(
-            "smallSquareYellow",
-            {
-              center: position,
-              innerRadius: 0,
-              outerRadius: satelliteRadius * 0.6,
-            },
-            { level: baseLevel },
-          );
-        },
-      );
-
-      /*const satelliteCountOuter = 32;
-      const satelliteRadiusOuter = 100;
-      const orbitRadiusOuter = 500 + 700 + satelliteRadius;
-
-      const satellitesOuter = Array.from({ length: satelliteCountOuter }, (_, index) => {
-        const angle = (index / satelliteCountOuter) * Math.PI * 2;
-        const position: SceneVector2 = {
-          x: center.x + Math.cos(angle) * orbitRadiusOuter,
-          y: center.y + Math.sin(angle) * orbitRadiusOuter,
-        };
-        return circleWithBricks(
-          "smallSquareGray",
-          {
-            center: position,
-            innerRadius: 0,
-            outerRadius: satelliteRadiusOuter,
-          },
-          { level: baseLevel + 0.25 }
-        );
-      });
-      */
-      return [
-        largeCircle,
-        largeYellowCircle,
-        ...satellites,
-        ...satellitesInner,
-      ];
-    },
-    playerUnits: [
-      {
-        type: "bluePentagon",
-        position: { x: 100, y: 100 },
-      },
-    ],
-    mapsRequired: { foundations: 1 },
-    maxLevel: 2,
-  },
-  turretRings: (() => {
-    const size: SceneSize = { width: 1400, height: 1400 };
-    const center: SceneVector2 = { x: size.width / 2, y: size.height / 2 };
-    const spawnPoint: SceneVector2 = { x: center.x, y: center.y - 600 };
-
-    // Позиції центрів окремих кіл (5-6 кіл)
-    const ringCenters: SceneVector2[] = [
-      { x: center.x - 300, y: center.y - 200 },
-      { x: center.x + 300, y: center.y - 200 },
-      { x: center.x - 300, y: center.y + 200 },
-      { x: center.x + 300, y: center.y + 200 },
-      { x: center.x, y: center.y - 400 },
-      { x: center.x, y: center.y + 400 },
-    ];
-
-    return {
-      name: "Turret Rings",
-      size,
-      icon: "ring_turrets.png",
-      unlockedBy: [
-        {
-          type: "map",
-          id: "initial",
-          level: 1,
-        },
-      ],
-      nodePosition: { x: 1, y: 2 },
-      maxLevel: 1,
-      spawnPoints: [spawnPoint],
-      bricks: ({ mapLevel }) => {
-        const level = Math.max(1, Math.floor(mapLevel));
-        const ringRadius = 150;
-        const brickSize = 24; // Розмір бріка smallSquareYellow
-        const ringThickness = 2 * brickSize; // Товщина 2 бріки = 48px
-        const innerRadius = ringRadius - ringThickness;
-        const outerRadius = ringRadius;
-
-        const rings: BrickShapeBlueprint[] = [];
-
-        // Генеруємо окремі кола піску (використовуємо smallSquareYellow, який дає sand)
-        ringCenters.forEach((ringCenter) => {
-          const ring = circleWithBricks(
-            "smallSquareYellow",
-            {
-              center: ringCenter,
-              innerRadius,
-              outerRadius,
-            },
-            { level },
-          );
-          rings.push(ring);
-        });
-
-        return rings;
-      },
-      enemies: ({ mapLevel }) => {
-        const level = Math.max(1, Math.floor(mapLevel));
-        const turrets: EnemySpawnData[] = [];
-
-        // Додаємо турель в центрі кожного кола
-        ringCenters.forEach((ringCenter) => {
-          turrets.push({
-            type: "turretEnemy",
-            level,
-            position: { ...ringCenter },
-          });
-        });
-
-        return turrets;
-      },
-      playerUnits: [
-        {
-          type: "bluePentagon",
-          position: { ...spawnPoint },
-        },
-      ],
-      mapsRequired: { initial: 1 },
-      resourceMultiplier: 2, // x2 бонус до ресурсів
-    } satisfies MapConfig;
-  })(),
-  sphinx: (() => {
-    const size: SceneSize = { width: 1400, height: 900 };
-    // Sphinx lying down, facing left
-    // Body center point
-    const bodyX = 700;
-    const bodyY = 550;
-    // Spawn point far from sphinx (top right corner)
-    const spawnPoint: SceneVector2 = { x: size.width - 150, y: 150 };
-
-    return {
-      name: "Sand Sphinx",
-      size,
-      icon: "sphynx.png",
-      spawnPoints: [spawnPoint],
-      nodePosition: { x: 2, y: 1 },
-      bricks: ({ mapLevel }) => {
-        const baseLevel = Math.max(0, Math.floor(mapLevel));
-        const sandLevel = baseLevel + 2;
-
-        // === BODY (lying lion, with curved back) ===
-        // Main body - wider polygon with curved appearance
-        const bodyMain = polygonWithBricks(
-          "smallSquareYellow",
-          {
-            vertices: [
-              { x: bodyX - 280, y: bodyY - 80 }, // front top
-              { x: bodyX - 100, y: bodyY - 100 }, // mid-front top (curved up)
-              { x: bodyX + 100, y: bodyY - 110 }, // mid-back top (highest point)
-              { x: bodyX + 280, y: bodyY - 80 }, // back top
-              { x: bodyX + 300, y: bodyY + 50 }, // back bottom
-              { x: bodyX - 280, y: bodyY + 50 }, // front bottom
-            ],
-          },
-          { level: sandLevel },
-        );
-
-        // Belly (adds roundness underneath)
-        const belly = circleWithBricks(
-          "smallSquareYellow",
-          {
-            center: { x: bodyX, y: bodyY + 20 },
-            innerRadius: 0,
-            outerRadius: 80,
-          },
-          { level: sandLevel },
-        );
-
-        // Haunches (back raised part - larger for folded hind legs)
-        const haunches = circleWithBricks(
-          "smallSquareYellow",
-          {
-            center: { x: bodyX + 180, y: bodyY - 50 },
-            innerRadius: 0,
-            outerRadius: 110,
-          },
-          { level: sandLevel },
-        );
-
-        // Back curve (adds more volume to the back)
-        const backCurve = circleWithBricks(
-          "smallSquareYellow",
-          {
-            center: { x: bodyX + 50, y: bodyY - 80 },
-            innerRadius: 0,
-            outerRadius: 70,
-          },
-          { level: sandLevel },
-        );
-
-        // === CHEST (raised front part) ===
-        const chest = polygonWithBricks(
-          "smallSquareYellow",
-          {
-            vertices: [
-              { x: bodyX - 340, y: bodyY - 140 },
-              { x: bodyX - 240, y: bodyY - 140 },
-              { x: bodyX - 220, y: bodyY - 40 },
-              { x: bodyX - 340, y: bodyY - 40 },
-            ],
-          },
-          { level: sandLevel },
-        );
-
-        // === NECK (shorter, connects to head) ===
-        const neck = polygonWithBricks(
-          "smallSquareYellow",
-          {
-            vertices: [
-              { x: bodyX - 380, y: bodyY - 200 },
-              { x: bodyX - 300, y: bodyY - 200 },
-              { x: bodyX - 260, y: bodyY - 130 },
-              { x: bodyX - 360, y: bodyY - 130 },
-            ],
-          },
-          { level: sandLevel },
-        );
-
-        // === HEAD (human profile facing left) ===
-        // Back of head (rounded)
-        const headBack = circleWithBricks(
-          "smallSquareYellow",
-          {
-            center: { x: bodyX - 360, y: bodyY - 260 },
-            innerRadius: 0,
-            outerRadius: 70,
-          },
-          { level: sandLevel },
-        );
-
-        // Face (polygon for profile - forehead, nose, chin)
-        const face = polygonWithBricks(
-          "smallSquareYellow",
-          {
-            vertices: [
-              { x: bodyX - 430, y: bodyY - 320 }, // forehead
-              { x: bodyX - 480, y: bodyY - 280 }, // nose tip
-              { x: bodyX - 470, y: bodyY - 250 }, // under nose
-              { x: bodyX - 480, y: bodyY - 220 }, // lips
-              { x: bodyX - 450, y: bodyY - 190 }, // chin
-              { x: bodyX - 400, y: bodyY - 200 }, // jaw
-              { x: bodyX - 380, y: bodyY - 260 }, // cheek
-              { x: bodyX - 400, y: bodyY - 310 }, // temple
-            ],
-          },
-          { level: sandLevel },
-        );
-
-        // Nemes headdress (flows down sides)
-        const nemesBack = polygonWithBricks(
-          "smallSquareYellow",
-          {
-            vertices: [
-              { x: bodyX - 320, y: bodyY - 300 },
-              { x: bodyX - 280, y: bodyY - 300 },
-              { x: bodyX - 260, y: bodyY - 180 },
-              { x: bodyX - 300, y: bodyY - 180 },
-            ],
-          },
-          { level: sandLevel },
-        );
-
-        // Crown/top of headdress
-        const crown = polygonWithBricks(
-          "smallSquareYellow",
-          {
-            vertices: [
-              { x: bodyX - 420, y: bodyY - 360 },
-              { x: bodyX - 340, y: bodyY - 360 },
-              { x: bodyX - 320, y: bodyY - 320 },
-              { x: bodyX - 430, y: bodyY - 320 },
-            ],
-          },
-          { level: sandLevel },
-        );
-
-        // === FRONT PAWS (extended forward, connected to body) ===
-        // Left front paw (closer, starts from chest)
-        const frontPawLeft = polygonWithBricks(
-          "smallSquareYellow",
-          {
-            vertices: [
-              { x: bodyX - 450, y: bodyY },
-              { x: bodyX - 300, y: bodyY },
-              { x: bodyX - 300, y: bodyY + 50 },
-              { x: bodyX - 450, y: bodyY + 50 },
-            ],
-          },
-          { level: sandLevel },
-        );
-
-        // Right front paw (slightly behind and lower)
-        const frontPawRight = polygonWithBricks(
-          "smallSquareYellow",
-          {
-            vertices: [
-              { x: bodyX - 400, y: bodyY + 40 },
-              { x: bodyX - 250, y: bodyY + 40 },
-              { x: bodyX - 250, y: bodyY + 90 },
-              { x: bodyX - 400, y: bodyY + 90 },
-            ],
-          },
-          { level: sandLevel },
-        );
-
-        // Paw ends (toes)
-        const pawEndLeft = circleWithBricks(
-          "smallSquareYellow",
-          {
-            center: { x: bodyX - 450, y: bodyY + 25 },
-            innerRadius: 0,
-            outerRadius: 30,
-          },
-          { level: sandLevel },
-        );
-
-        const pawEndRight = circleWithBricks(
-          "smallSquareYellow",
-          {
-            center: { x: bodyX - 400, y: bodyY + 65 },
-            innerRadius: 0,
-            outerRadius: 28,
-          },
-          { level: sandLevel },
-        );
-
-        // === HIND LEG (left, visible from side - extended like front paws) ===
-        // Upper thigh (connects to haunches)
-        const hindThighLeft = polygonWithBricks(
-          "smallSquareYellow",
-          {
-            vertices: [
-              { x: bodyX + 80, y: bodyY - 20 },
-              { x: bodyX + 160, y: bodyY - 20 },
-              { x: bodyX + 180, y: bodyY + 40 },
-              { x: bodyX + 100, y: bodyY + 40 },
-            ],
-          },
-          { level: sandLevel },
-        );
-
-        // Lower leg (extends forward like front paws)
-        const hindLegLeft = polygonWithBricks(
-          "smallSquareYellow",
-          {
-            vertices: [
-              { x: bodyX + 60, y: bodyY + 30 },
-              { x: bodyX + 180, y: bodyY + 30 },
-              { x: bodyX + 180, y: bodyY + 80 },
-              { x: bodyX + 60, y: bodyY + 80 },
-            ],
-          },
-          { level: sandLevel },
-        );
-
-        // Hind paw (like front paws)
-        const hindPawLeft = circleWithBricks(
-          "smallSquareYellow",
-          {
-            center: { x: bodyX + 60, y: bodyY + 55 },
-            innerRadius: 0,
-            outerRadius: 32,
-          },
-          { level: sandLevel },
-        );
-
-        // === TAIL (more horizontal, along the ground) ===
-        const tail1 = circleWithBricks(
-          "smallSquareYellow",
-          {
-            center: { x: bodyX + 330, y: bodyY + 30 },
-            innerRadius: 0,
-            outerRadius: 35,
-          },
-          { level: sandLevel },
-        );
-
-        const tail2 = circleWithBricks(
-          "smallSquareYellow",
-          {
-            center: { x: bodyX + 380, y: bodyY + 20 },
-            innerRadius: 0,
-            outerRadius: 30,
-          },
-          { level: sandLevel },
-        );
-
-        const tail3 = circleWithBricks(
-          "smallSquareYellow",
-          {
-            center: { x: bodyX + 420, y: bodyY + 15 },
-            innerRadius: 0,
-            outerRadius: 25,
-          },
-          { level: sandLevel },
-        );
-
-        // === SAND BASE ===
-        const sandBase = polygonWithBricks(
-          "smallSquareYellow",
-          {
-            vertices: [
-              { x: 150, y: bodyY + 120 },
-              { x: size.width - 150, y: bodyY + 120 },
-              { x: size.width - 130, y: bodyY + 160 },
-              { x: 130, y: bodyY + 160 },
-            ],
-          },
-          { level: sandLevel },
-        );
-
-        return [
-          sandBase,
-          bodyMain,
-          belly,
-          backCurve,
-          haunches,
-          chest,
-          neck,
-          headBack,
-          face,
-          nemesBack,
-          crown,
-          frontPawLeft,
-          frontPawRight,
-          pawEndLeft,
-          pawEndRight,
-          hindThighLeft,
-          hindLegLeft,
-          hindPawLeft,
-          tail1,
-          tail2,
-          tail3,
-        ];
-      },
-      playerUnits: [
-        {
-          type: "bluePentagon",
-          position: { ...spawnPoint },
-        },
-      ],
-      unlockedBy: [
-        {
-          type: "map",
-          id: "initial",
-          level: 2,
-        },
-      ],
-      mapsRequired: { initial: 2 },
-      maxLevel: 1,
-    } satisfies MapConfig;
-  })(),
-  thicket: (() => {
-    const size: SceneSize = { width: 1000, height: 1000 };
-    const sandHeight = 60;
-    const createRectangle = (
-      x: number,
-      y: number,
-      width: number,
-      height: number,
-    ): SceneVector2[] => [
-      { x, y },
-      { x: x + width, y },
-      { x: x + width, y: y + height },
-      { x, y: y + height },
-    ];
-    const cactusStart = size.height - sandHeight - 200;
-    const sandVertices = createRectangle(
-      0,
-      size.height - sandHeight - 200,
-      size.width,
-      sandHeight,
-    );
-    const bushClusters: readonly { center: SceneVector2; radius: number }[] = [
-      { center: { x: 320, y: 560 }, radius: 110 },
-      { center: { x: 500, y: 440 }, radius: 100 },
-      { center: { x: 720, y: 500 }, radius: 105 },
-      { center: { x: 880, y: 360 }, radius: 115 },
-      { center: { x: 620, y: 620 }, radius: 90 },
-    ];
-
-    return {
-      name: "Overgrown Thicket",
-      size,
-      icon: "thicket.png",
-      spawnPoints: [{ x: 50, y: size.height - sandHeight - 420 }],
-      nodePosition: { x: 2, y: 3 },
-      bricks: ({ mapLevel }) => {
-        const baseLevel = Math.max(0, Math.floor(mapLevel));
-        const sandLevel = baseLevel + 1;
-        const organicLevel = baseLevel;
-
-        const sandBank = polygonWithBricks(
-          "smallSquareYellow",
-          {
-            vertices: sandVertices,
-          },
-          { level: sandLevel },
-        );
-
-        /*const bushes = bushClusters.map((cluster) =>
-          circleWithBricks(
-            "smallOrganic",
-            {
-              center: cluster.center,
-              innerRadius: 0,
-              outerRadius: cluster.radius,
-            },
-            { level: organicLevel },
-          ),
-        );*/
-
-        const cactus = [];
-
-        cactus.push(
-          polygonWithBricks(
-            "smallOrganic",
-            {
-              vertices: createRectangle(420, cactusStart - 360, 120, 360),
-            },
-            { level: organicLevel },
-          )
-        );
-
-        cactus.push(
-          circleWithBricks(
-            "smallOrganic",
-            {
-              center: { x: 480, y: cactusStart - 360 },
-              innerRadius: 0,
-              outerRadius: 60,
-            },
-            { level: organicLevel },
-          )
-        );
-
-        cactus.push(
-          polygonWithBricks(
-            "smallOrganic",
-            {
-              vertices: createRectangle(280, cactusStart - 460, 90, 310),
-            },
-            { level: organicLevel },
-          )
-        );
-
-        cactus.push(
-          circleWithBricks(
-            "smallOrganic",
-            {
-              center: { x: 320, y: cactusStart - 460 },
-              innerRadius: 0,
-              outerRadius: 45,
-            },
-            { level: organicLevel },
-          )
-        );
-
-        cactus.push(
-          circleWithBricks(
-            "smallOrganic",
-            {
-              center: { x: 390, y: cactusStart - 160 },
-              innerRadius: 0,
-              outerRadius: 120,
-            },
-            { level: organicLevel },
-          )
-        );
-
-        cactus.push(
-          polygonWithBricks(
-            "smallOrganic",
-            {
-              vertices: createRectangle(640, cactusStart - 360, 90, 210),
-            },
-            { level: organicLevel },
-          )
-        );
-
-        cactus.push(
-          circleWithBricks(
-            "smallOrganic",
-            {
-              center: { x: 680, y: cactusStart - 360 },
-              innerRadius: 0,
-              outerRadius: 45,
-            },
-            { level: organicLevel },
-          )
-        );
-
-        cactus.push(
-          circleWithBricks(
-            "smallOrganic",
-            {
-              center: { x: 600, y: cactusStart - 120 },
-              innerRadius: 0,
-              outerRadius: 120,
-            },
-            { level: organicLevel },
-          )
-        );
-
-
-        return [sandBank, ...cactus];
-      },
-      playerUnits: [
-        {
-          type: "bluePentagon",
-          position: { x: 50, y: size.height - sandHeight - 220 },
-        },
-      ],
-      unlockedBy: [
-        {
-          type: "map",
-          id: "initial",
-          level: 1,
-        },
-      ],
-      mapsRequired: { initial: 1 },
-      maxLevel: 1,
-    } satisfies MapConfig;
-  })(),
-  oldForge: (() => {
-    const size: SceneSize = { width: 1000, height: 1000 };
-    const center: SceneVector2 = { x: size.width / 2, y: size.height / 2 };
-    const outerSize = 700;
-    const cavitySize = 500;
-
-    const enemySpawnPosition: SceneVector2 = {
-      x: center.x,
-      y: center.y + outerSize / 2 + 20,
-    };
-
-    return {
-      name: "Old Forge",
-      size,
-      icon: "forge.png",
-      spawnPoints: [{ x: center.x, y: center.y - outerSize / 2 + 80 }],
-      /*enemySpawnPoints: [
-        {
-          position: enemySpawnPosition,
-          spawnRate: 0.2, // 1 ворог на 5 секунд (1/5 = 0.2)
-          enemyTypes: [
-            { type: "tankEnemy", weight: 1.0 },
-            { type: "fastEnemy", weight: 1.0 },
-          ],
-          maxConcurrent: 10,
-        },
-      ],*/
-      nodePosition: { x: 3, y: 2 },
-      bricks: ({ mapLevel }) => {
-        const baseLevel = Math.max(0, Math.floor(mapLevel));
-        const walkwayLevel = baseLevel + 1;
-        const ironThickness = getBrickConfig("smallIron").size.width;
-        const innerRingSize = Math.max(cavitySize - ironThickness * 8, 0);
-
-        const forgeFloor = squareWithBricks(
-          "smallSquareGray",
-          {
-            center,
-            size: outerSize,
-            innerSize: cavitySize,
-          },
-          { level: walkwayLevel },
-        );
-
-        const ironLining = squareWithBricks(
-          "smallIron",
-          {
-            center,
-            size: cavitySize,
-            innerSize: innerRingSize > 0 ? innerRingSize : undefined,
-          },
-          { level: baseLevel },
-        );
-
-        return [forgeFloor, ironLining];
-      },
-      playerUnits: [
-        {
-          type: "bluePentagon",
-          position: { x: center.x, y: center.y - outerSize / 2 + 80 },
-        },
-      ],
-      unlockedBy: [
-        {
-          type: "map",
-          id: "initial",
-          level: 1,
-        },
-      ],
-      mapsRequired: { initial: 1 },
-      maxLevel: 1,
-    } satisfies MapConfig;
-  })(),
-  spruce: (() => {
-    const size: SceneSize = { width: 1500, height: 1600 };
-    const createRectangle = (
-      x: number,
-      y: number,
-      width: number,
-      height: number,
-    ): SceneVector2[] => [
-      { x, y },
-      { x: x + width, y },
-      { x: x + width, y: y + height },
-      { x, y: y + height },
-    ];
-    const createTriangle = (
-      baseCenter: SceneVector2,
-      width: number,
-      height: number,
-    ): SceneVector2[] => [
-      { x: baseCenter.x, y: baseCenter.y - height },
-      { x: baseCenter.x + width / 2, y: baseCenter.y },
-      { x: baseCenter.x - width / 2, y: baseCenter.y },
-    ];
-
-    const treeConfigs: readonly { base: SceneVector2; scale: number }[] = [
-      { base: { x: 350, y: 1200 }, scale: 1 },
-      { base: { x: 650, y: 1100 }, scale: 0.95 },
-      { base: { x: 950, y: 1250 }, scale: 1.1 },
-      { base: { x: 1230, y: 1150 }, scale: 0.9 },
-      { base: { x: 500, y: 900 }, scale: 0.85 },
-      { base: { x: 1050, y: 880 }, scale: 0.9 },
-    ];
-
-    const spawnPoint: SceneVector2 = { x: 200, y: 1300 };
-
-    return {
-      name: "Forest",
-      size,
-      spawnPoints: [spawnPoint],
-      nodePosition: { x: 3, y: 4 },
-      icon: "spruce.png",
-      bricks: ({ mapLevel }) => {
-        const baseLevel = Math.max(0, Math.floor(mapLevel));
-        const canopyLevel = baseLevel + 1;
-        const trunkLevel = baseLevel;
-
-        const trees = treeConfigs.flatMap((tree) => {
-          const trunkHeight = 180 * tree.scale;
-          const trunkWidth = 60 * tree.scale;
-          const trunkBottomY = tree.base.y;
-          const trunkTopY = trunkBottomY - trunkHeight;
-
-          const trunk = polygonWithBricks(
-            "smallWood",
-            {
-              vertices: createRectangle(
-                tree.base.x - trunkWidth / 2,
-                trunkTopY,
-                trunkWidth,
-                trunkHeight,
-              ),
-            },
-            { level: trunkLevel },
-          );
-
-          const canopyLayers = [
-            { width: 320, height: 260, offset: 20 },
-            { width: 260, height: 220, offset: 120 },
-            { width: 190, height: 180, offset: 210 },
-          ];
-
-          const canopy = canopyLayers.map((layer) => {
-            const baseCenter: SceneVector2 = {
-              x: tree.base.x,
-              y: trunkTopY + layer.offset * tree.scale,
-            };
-            return polygonWithBricks(
-              "smallOrganic",
-              {
-                vertices: createTriangle(
-                  baseCenter,
-                  layer.width * tree.scale,
-                  layer.height * tree.scale,
-                ),
-              },
-              { level: canopyLevel },
-            );
-          });
-
-          return [trunk, ...canopy];
-        });
-
-        return trees;
-      },
-      playerUnits: [
-        {
-          type: "bluePentagon",
-          position: { ...spawnPoint },
-        },
-      ],
-      unlockedBy: [
-        {
-          type: "map",
-          id: "thicket",
-          level: 1,
-        },
-      ],
-      mapsRequired: { thicket: 1 },
-      maxLevel: 1,
-    } satisfies MapConfig;
-  })(),
-  deadOak: (() => {
-    const size: SceneSize = { width: 1400, height: 1200 };
-    const centerX = size.width / 2;
-    const groundY = size.height - 180;
-    const spawnPoint: SceneVector2 = { x: 140, y: 140 };
-
-    return {
-      name: "Dead Oak",
-      size,
-      spawnPoints: [spawnPoint],
-      nodePosition: { x: 3, y: 5 },
-      icon: "dead_oak.png",
-      bricks: ({ mapLevel }) => {
-        const baseLevel = Math.max(0, Math.floor(mapLevel));
-        const woodLevel = baseLevel + 1;
-        const trunkTopY = groundY - 520;
-        const trunkWidthBase = 170;
-        const trunkWidthTop = 90;
-
-        const trunkOutline = [
-          {
-            start: { x: centerX - trunkWidthBase / 2, y: groundY },
-            control1: { x: centerX - trunkWidthBase / 2 - 20, y: groundY - 140 },
-            control2: { x: centerX - trunkWidthTop / 2 - 30, y: trunkTopY + 160 },
-            end: { x: centerX - trunkWidthTop / 2, y: trunkTopY },
-          },
-          {
-            start: { x: centerX - trunkWidthTop / 2, y: trunkTopY },
-            control1: { x: centerX - trunkWidthTop / 4, y: trunkTopY - 20 },
-            control2: { x: centerX + trunkWidthTop / 4, y: trunkTopY - 20 },
-            end: { x: centerX + trunkWidthTop / 2, y: trunkTopY },
-          },
-          {
-            start: { x: centerX + trunkWidthTop / 2, y: trunkTopY },
-            control1: { x: centerX + trunkWidthTop / 2 + 30, y: trunkTopY + 160 },
-            control2: { x: centerX + trunkWidthBase / 2 + 20, y: groundY - 140 },
-            end: { x: centerX + trunkWidthBase / 2, y: groundY },
-          },
-          {
-            start: { x: centerX + trunkWidthBase / 2, y: groundY },
-            control1: { x: centerX + trunkWidthBase / 4, y: groundY + 20 },
-            control2: { x: centerX - trunkWidthBase / 4, y: groundY + 20 },
-            end: { x: centerX - trunkWidthBase / 2, y: groundY },
-          },
-        ] as const;
-
-        const trunk = bezierPolygonWithBricks(
-          "smallWood",
-          {
-            outline: trunkOutline,
-            spacing: 28,
-            sampleStep: 14,
-            alignToEdge: true,
-          },
-          { level: woodLevel },
-        );
-
-        const leftMainSegments = [
-          {
-            start: { x: centerX - 40, y: trunkTopY + 240 },
-            control1: { x: centerX - 160, y: trunkTopY + 160 },
-            control2: { x: centerX - 300, y: trunkTopY + 100 },
-            end: { x: centerX - 360, y: trunkTopY + 80 },
-          },
-        ] as const;
-
-        const rightMainSegments = [
-          {
-            start: { x: centerX + 40, y: trunkTopY + 240 },
-            control1: { x: centerX + 160, y: trunkTopY + 160 },
-            control2: { x: centerX + 300, y: trunkTopY + 100 },
-            end: { x: centerX + 360, y: trunkTopY + 80 },
-          }
-        ] as const;
-
-        const leftMainBottomSegments = [
-          {
-            start: { x: centerX - 360, y: trunkTopY + 90 },
-            control1: { x: centerX - 410, y: trunkTopY + 90 },
-            control2: { x: centerX - 490, y: trunkTopY + 130 },
-            end: { x: centerX - 610, y: trunkTopY + 150 },
-          },
-        ] as const;
-
-        const rightMainBottomSegments = [
-          {
-            start: { x: centerX + 360, y: trunkTopY + 90 },
-            control1: { x: centerX + 410, y: trunkTopY + 90 },
-            control2: { x: centerX + 490, y: trunkTopY + 130 },
-            end: { x: centerX + 610, y: trunkTopY + 150 },
-          },
-        ] as const;
-
-        const leftMainTopSegments = [
-          {
-            start: { x: centerX - 360, y: trunkTopY + 70 },
-            control1: { x: centerX - 410, y: trunkTopY + 40 },
-            control2: { x: centerX - 490, y: trunkTopY + 40 },
-            end: { x: centerX - 610, y: trunkTopY - 50 },
-          },
-        ] as const;
-
-        const rightMainTopSegments = [
-          {
-            start: { x: centerX + 360, y: trunkTopY + 70 },
-            control1: { x: centerX + 410, y: trunkTopY + 40 },
-            control2: { x: centerX + 490, y: trunkTopY + 40 },
-            end: { x: centerX + 610, y: trunkTopY - 50 },
-          },
-        ] as const;
-
-        const leftUpperBranchSegments =[
-          {
-            start: { x: centerX - 10, y: trunkTopY - 10 },
-            control1: { x: centerX - 210, y: trunkTopY - 140 },
-            control2: { x: centerX - 390, y: trunkTopY - 170 },
-            end: { x: centerX - 550, y: trunkTopY - 200 },
-          },
-        ] as const;
-
-        const rightUpperBranchSegments = [
-          {
-            start: { x: centerX + 10, y: trunkTopY - 10 },
-            control1: { x: centerX + 210, y: trunkTopY - 140 },
-            control2: { x: centerX + 390, y: trunkTopY - 170 },
-            end: { x: centerX + 550, y: trunkTopY - 200 },
-          },
-        ] as const;
-
-        const leftMainBranch = bezierCurveWithBricks(
-          "smallWood",
-          {
-            segments: leftMainSegments,
-            spacing: 26,
-            thickness: 60,
-          },
-          { level: woodLevel },
-        );
-
-        const rightMainBranch = bezierCurveWithBricks(
-          "smallWood",
-          {
-            segments: rightMainSegments,
-            spacing: 26,
-            thickness: 60,
-          },
-          { level: woodLevel },
-        );
-
-        const leftMainBranchBottom = bezierCurveWithBricks(
-          "smallWood",
-          {
-            segments: leftMainBottomSegments,
-            spacing: 26,
-            thickness: 40,
-          },
-          { level: woodLevel },
-        );
-
-        const rightMainBranchBottom = bezierCurveWithBricks(
-          "smallWood",
-          {
-            segments: rightMainBottomSegments,
-            spacing: 26,
-            thickness: 40,
-          },
-          { level: woodLevel },
-        );
-
-        const leftMainBranchTop = bezierCurveWithBricks(
-          "smallWood",
-          {
-            segments: leftMainTopSegments,
-            spacing: 26,
-            thickness: 40,
-          },
-          { level: woodLevel },
-        );
-        const rightMainBranchTop = bezierCurveWithBricks(
-          "smallWood",
-          {
-            segments: rightMainTopSegments,
-            spacing: 26,
-            thickness: 40,
-          },
-          { level: woodLevel },
-        );
-
-        const leftUpperBranch = bezierCurveWithBricks(
-          "smallWood",
-          {
-            segments: leftUpperBranchSegments,
-            spacing: 26,
-            thickness: 30,
-          },
-          { level: woodLevel },
-        );
-
-        const rightUpperBranch = bezierCurveWithBricks(
-          "smallWood",
-          {
-            segments: rightUpperBranchSegments,
-            spacing: 26,
-            thickness: 30,
-          },
-          { level: woodLevel },
-        );
-
-        const rootsOutline = [
-          {
-            start: { x: centerX - 180, y: groundY },
-            control1: { x: centerX - 160, y: groundY + 40 },
-            control2: { x: centerX - 60, y: groundY + 60 },
-            end: { x: centerX, y: groundY + 50 },
-          },
-          {
-            start: { x: centerX, y: groundY + 50 },
-            control1: { x: centerX + 60, y: groundY + 60 },
-            control2: { x: centerX + 160, y: groundY + 40 },
-            end: { x: centerX + 180, y: groundY },
-          },
-          {
-            start: { x: centerX + 180, y: groundY },
-            control1: { x: centerX + 140, y: groundY - 10 },
-            control2: { x: centerX + 80, y: groundY - 10 },
-            end: { x: centerX, y: groundY - 10 },
-          },
-          {
-            start: { x: centerX, y: groundY - 10 },
-            control1: { x: centerX - 80, y: groundY - 10 },
-            control2: { x: centerX - 140, y: groundY - 10 },
-            end: { x: centerX - 180, y: groundY },
-          },
-        ] as const;
-
-        const roots = bezierPolygonWithBricks(
-          "smallWood",
-          {
-            outline: rootsOutline,
-            spacing: 26,
-            sampleStep: 12,
-            alignToEdge: true,
-          },
-          { level: woodLevel },
-        );
-
-        return [
-          trunk,
-          leftMainBranch,
-          rightMainBranch,
-          leftMainBranchBottom,
-          rightMainBranchBottom,
-          leftMainBranchTop,
-          rightMainBranchTop,
-          leftUpperBranch,
-          rightUpperBranch,
-          roots,
-         bezierCurveWithBricks(
-            "smallWood",
-            {
-              segments: [
-                {
-                  start: { x: centerX - 10, y: trunkTopY - 10 },
-                  control1: { x: centerX - 30, y: trunkTopY - 70 },
-                  control2: { x: centerX - 50, y: trunkTopY - 110 },
-                  end: { x: centerX - 50, y: trunkTopY - 140 },
-                },
-                {
-                  start: { x: centerX - 50, y: trunkTopY - 140 },
-                  control1: { x: centerX - 90, y: trunkTopY - 160 },
-                  control2: { x: centerX - 150, y: trunkTopY - 190 },
-                  end: { x: centerX - 210, y: trunkTopY - 230 },
-                },
-                {
-                  start: { x: centerX - 210, y: trunkTopY - 230 },
-                  control1: { x: centerX - 250, y: trunkTopY - 250 },
-                  control2: { x: centerX - 280, y: trunkTopY - 280 },
-                  end: { x: centerX - 310, y: trunkTopY - 360 },
-                },
-                
-              ],
-              spacing: 26,
-              thickness: 28,
-            },
-            { level: woodLevel },
-          ),
-          bezierCurveWithBricks(
-            "smallWood",
-            {
-              segments: [
-                {
-                  start: { x: centerX + 10, y: trunkTopY - 10 },
-                  control1: { x: centerX + 30, y: trunkTopY - 70 },
-                  control2: { x: centerX + 50, y: trunkTopY - 110 },
-                  end: { x: centerX + 50, y: trunkTopY - 140 },
-                },
-                {
-                  start: { x: centerX + 50, y: trunkTopY - 140 },
-                  control1: { x: centerX + 90, y: trunkTopY - 160 },
-                  control2: { x: centerX + 150, y: trunkTopY - 190 },
-                  end: { x: centerX + 210, y: trunkTopY - 230 },
-                },
-                {
-                  start: { x: centerX + 210, y: trunkTopY - 230 },
-                  control1: { x: centerX + 250, y: trunkTopY - 250 },
-                  control2: { x: centerX + 280, y: trunkTopY - 280 },
-                  end: { x: centerX + 310, y: trunkTopY - 360 },
-                },
-                
-              ],
-              spacing: 26,
-              thickness: 28,
-            },
-            { level: woodLevel },
-          ),
-          bezierCurveWithBricks(
-            "smallWood",
-            {
-              segments: [
-                {
-                  start: { x: centerX - 50, y: trunkTopY - 140 },
-                  control1: { x: centerX - 60, y: trunkTopY - 210 },
-                  control2: { x: centerX - 120, y: trunkTopY - 240 },
-                  end: { x: centerX - 130, y: trunkTopY - 390 },
-                },
-              ],
-              spacing: 26,
-              thickness: 20,
-            },
-            { level: woodLevel },
-          ),
-          bezierCurveWithBricks(
-            "smallWood",
-            {
-              segments: [
-                {
-                  start: { x: centerX + 50, y: trunkTopY - 140 },
-                  control1: { x: centerX + 60, y: trunkTopY - 210 },
-                  control2: { x: centerX + 120, y: trunkTopY - 240 },
-                  end: { x: centerX + 130, y: trunkTopY - 390 },
-                },
-              ],
-              spacing: 26,
-              thickness: 20,
-            },
-            { level: woodLevel },
-          ),
-          bezierCurveWithBricks(
-            "smallWood",
-            {
-              segments: [
-                {
-                  start: { x: centerX + 50, y: trunkTopY - 140 },
-                  control1: { x: centerX + 10, y: trunkTopY - 210 },
-                  control2: { x: centerX - 30, y: trunkTopY - 240 },
-                  end: { x: centerX - 10, y: trunkTopY - 390 },
-                },
-              ],
-              spacing: 26,
-              thickness: 20,
-            },
-            { level: woodLevel },
-          ),
-          bezierCurveWithBricks(
-            "smallWood",
-            {
-              segments: [
-                {
-                  start: { x: centerX + 150, y: trunkTopY + 210 },
-                  control1: { x: centerX + 190, y: trunkTopY + 240 },
-                  control2: { x: centerX + 320, y: trunkTopY + 270 },
-                  end: { x: centerX + 430, y: trunkTopY + 300 },
-                },
-              ],
-              spacing: 26,
-              thickness: 20,
-            },
-            { level: woodLevel },
-          ),
-          bezierCurveWithBricks(
-            "smallWood",
-            {
-              segments: [
-                {
-                  start: { x: centerX - 150, y: trunkTopY + 210 },
-                  control1: { x: centerX - 190, y: trunkTopY + 240 },
-                  control2: { x: centerX - 320, y: trunkTopY + 270 },
-                  end: { x: centerX - 430, y: trunkTopY + 300 },
-                },
-              ],
-              spacing: 26,
-              thickness: 20,
-            },
-            { level: woodLevel },
-          ),
-          bezierCurveWithBricks(
-            "smallWood",
-            {
-              segments: [
-                {
-                  start: { x: centerX + 150, y: trunkTopY - 70 },
-                  control1: { x: centerX + 250, y: trunkTopY - 90 },
-                  control2: { x: centerX + 320, y: trunkTopY - 60 },
-                  end: { x: centerX + 460, y: trunkTopY - 80 },
-                },
-              ],
-              spacing: 26,
-              thickness: 20,
-            },
-            { level: woodLevel },
-          ),
-          bezierCurveWithBricks(
-            "smallWood",
-            {
-              segments: [
-                {
-                  start: { x: centerX - 150, y: trunkTopY - 70 },
-                  control1: { x: centerX - 250, y: trunkTopY - 90 },
-                  control2: { x: centerX - 320, y: trunkTopY - 60 },
-                  end: { x: centerX - 460, y: trunkTopY - 80 },
-                },
-              ],
-              spacing: 26,
-              thickness: 20,
-            },
-            { level: woodLevel },
-          ),
-        ];
-      },
-      playerUnits: [
-        {
-          type: "bluePentagon",
-          position: { ...spawnPoint },
-        },
-      ],
-      unlockedBy: [
-        {
-          type: "map",
-          id: "spruce",
-          level: 1,
-        },
-      ],
-      mapsRequired: { spruce: 1 },
-      maxLevel: 2,
-    } satisfies MapConfig;
-  })(),
-  stoneCottage: (() => {
-    const size: SceneSize = { width: 1300, height: 1400 };
-    const center: SceneVector2 = { x: size.width / 2, y: (size.height-200) / 2 };
-    const spawnPoint: SceneVector2 = { x: center.x, y: size.height - 360 };
-
-    const createRectangle = (
-      x: number,
-      y: number,
-      width: number,
-      height: number,
-    ): SceneVector2[] => [
-      { x, y },
-      { x: x + width, y },
-      { x: x + width, y: y + height },
-      { x, y: y + height },
-    ];
-
-    const enemySpawnPosition: SceneVector2 = { x: size.width - 200, y: 200 };
-
-    return {
-      name: "Stone Cottage",
-      size,
-      spawnPoints: [spawnPoint],
-      enemySpawnPoints: [
-        {
-          position: enemySpawnPosition,
-          spawnRate: 0.1, // 1 ворог на 5 секунд (1/5 = 0.2)
-          enemyTypes: [{ type: "spectreEnemy", weight: 1.0 }],
-          maxConcurrent: 5,
-        },
-      ],
-      nodePosition: { x: 1, y: 4 },
-      icon: "cottage.png",
-      bricks: ({ mapLevel }) => {
-        const baseLevel = Math.max(0, Math.floor(mapLevel));
-        const stoneLevel = baseLevel + 3;
-        const ironLevel = baseLevel + 1;
-        const organicLevel = baseLevel + 1;
-
-        const walls = polygonWithBricks(
-          "smallSquareGray",
-          {
-            vertices: createRectangle(center.x - 260, center.y - 220, 520, 320),
-            holes: [createRectangle(center.x - 80, center.y - 150, 120, 160)],
-          },
-          { level: stoneLevel },
-        );
-
-        const roof = polygonWithBricks(
-          "smallIron",
-          {
-            vertices: [
-              { x: center.x - 300, y: center.y - 220 },
-              { x: center.x, y: center.y - 360 },
-              { x: center.x + 300, y: center.y - 220 },
-            ],
-          },
-          { level: ironLevel },
-        );
-
-        const doorFrame = polygonWithBricks(
-          "smallIron",
-          {
-            vertices: createRectangle(center.x - 60, center.y + 120, 120, 120),
-          },
-          { level: ironLevel },
-        );
-
-        const chimney = polygonWithBricks(
-          "smallSquareGray",
-          {
-            vertices: createRectangle(center.x + 140, center.y - 340, 70, 180),
-          },
-          { level: stoneLevel },
-        );
-
-        const bushes = [
-          circleWithBricks(
-            "smallOrganic",
-            {
-              center: { x: center.x - 280, y: center.y + 200 },
-              innerRadius: 0,
-              outerRadius: 90,
-            },
-            { level: organicLevel },
-          ),
-          circleWithBricks(
-            "smallOrganic",
-            {
-              center: { x: center.x + 280, y: center.y + 200 },
-              innerRadius: 0,
-              outerRadius: 100,
-            },
-            { level: organicLevel },
-          ),
-          circleWithBricks(
-            "smallOrganic",
-            {
-              center: { x: center.x + 320, y: center.y + 20 },
-              innerRadius: 0,
-              outerRadius: 100,
-            },
-            { level: organicLevel },
-          ),
-          circleWithBricks(
-            "smallOrganic",
-            {
-              center: { x: center.x + 400, y: center.y - 90 },
-              innerRadius: 0,
-              outerRadius: 100,
-            },
-            { level: organicLevel },
-          ),
-          circleWithBricks(
-            "smallOrganic",
-            {
-              center: { x: center.x + 390, y: center.y + 140 },
-              innerRadius: 0,
-              outerRadius: 100,
-            },
-            { level: organicLevel },
-          ),
-          circleWithBricks(
-            "smallOrganic",
-            {
-              center: { x: center.x - 200, y: center.y + 260 },
-              innerRadius: 0,
-              outerRadius: 70,
-            },
-            { level: organicLevel },
-          ),
-          circleWithBricks(
-            "smallOrganic",
-            {
-              center: { x: center.x - 130, y: center.y + 200 },
-              innerRadius: 0,
-              outerRadius: 70,
-            },
-            { level: organicLevel },
-          ),
-        ];
-
-        const courtyard = polygonWithBricks(
-          "smallSquareGray",
-          {
-            vertices: createRectangle(center.x - 140, center.y + 200, 280, 100),
-          },
-          { level: stoneLevel - 1 },
-        );
-
-        return [walls, roof, doorFrame, chimney, courtyard, ...bushes];
-      },
-      playerUnits: [
-        {
-          type: "bluePentagon",
-          position: { ...spawnPoint },
-        },
-      ],
-      unlockedBy: [
-        {
-          type: "map",
-          id: "thicket",
-          level: 1,
-        },
-      ],
-      mapsRequired: { thicket: 1 },
-      maxLevel: 2,
-    } satisfies MapConfig;
-  })(),
-  bezierGrove: (() => {
-    const size: SceneSize = { width: 1500, height: 1500 };
-    const center: SceneVector2 = { x: size.width / 2, y: (size.height - 120) / 2 };
-    const spawnPoint: SceneVector2 = { x: center.x - 380, y: size.height - 460 };
-
-    const vineSegmentsRight = [
-      {
-        start: { x: center.x, y: center.y + 320 },
-        control1: { x: center.x - 60, y: center.y + 80 },
-        control2: { x: center.x + 280, y: center.y - 40 },
-        end: { x: center.x + 420, y: center.y - 140 },
-      },
-      {
-        start: { x: center.x + 420, y: center.y - 120 },
-        control1: { x: center.x + 540, y: center.y - 180 },
-        control2: { x: center.x + 600, y: center.y - 300 },
-        end: { x: center.x + 680, y: center.y - 360 },
-      },
-    ] as const;
-
-    const vineSegmentsLeft = [
-      {
-        start: { x: center.x, y: center.y + 320 },
-        control1: { x: center.x + 60, y: center.y + 80 },
-        control2: { x: center.x - 280, y: center.y - 40 },
-        end: { x: center.x - 420, y: center.y - 110 },
-      },
-      {
-        start: { x: center.x - 420, y: center.y - 110 },
-        control1: { x: center.x - 540, y: center.y - 200 },
-        control2: { x: center.x - 600, y: center.y - 270 },
-        end: { x: center.x - 680, y: center.y - 320 },
-      },
-    ] as const;
-
-    const vineSegmentsLeftCenter = [
-      {
-        start: { x: center.x, y: center.y + 320 },
-        control1: { x: center.x + 160, y: center.y + 80 },
-        control2: { x: center.x + 280, y: center.y - 40 },
-        end: { x: center.x + 120, y: center.y - 110 },
-      },
-      {
-        start: { x: center.x + 120, y: center.y - 110 },
-        control1: { x: center.x + 40, y: center.y - 200 },
-        control2: { x: center.x - 60, y: center.y - 270 },
-        end: { x: center.x + 80, y: center.y - 320 },
-      },
-    ] as const;
-
-
-
-    const baseLeafOutline = [
-      {
-        start: { x: -120, y: 0 },
-        control1: { x: -80, y: -80 },
-        control2: { x: 20, y: -120 },
-        end: { x: 120, y: 0 },
-      },
-      {
-        start: { x: 120, y: 0 },
-        control1: { x: 20, y: 120 },
-        control2: { x: -80, y: 80 },
-        end: { x: -120, y: 0 },
-      },
-    ] as const;
-
-    const leafOutlineLeft = transformBezierOutline(baseLeafOutline, {
-      position: { x: center.x - 240, y: center.y + 140 },
-      scale: 0.9,
-      rotation: -Math.PI / 12,
-    });
-
-    const leafOutlineMid = transformBezierOutline(baseLeafOutline, {
-      position: { x: center.x + 40, y: center.y + 220 },
-      scale: 0.7,
-      rotation: Math.PI / 9,
-    });
-
-    const leafOutlineRight = transformBezierOutline(baseLeafOutline, {
-      position: { x: center.x + 260, y: center.y - 40 },
-      scale: 1.1,
-      rotation: (5 * Math.PI) / 18,
-    });
-
-    return {
-      name: "Dangerous Bushes",
-      size,
-      spawnPoints: [spawnPoint],
-      nodePosition: { x: 1, y: 5 },
-      icon: "dangerous_bushes.png",
-      lockedForDemo: true,
-      bricks: ({ mapLevel }) => {
-        const baseLevel = Math.max(0, Math.floor(mapLevel));
-        const woodLevel = baseLevel + 2;
-        const organicLevel = baseLevel + 3;
-
-        const mainVineRight = bezierCurveWithBricks(
-          "smallWood",
-          {
-            segments: vineSegmentsRight,
-            spacing: 26,
-            rotationOffset: Math.PI / 2,
-          },
-          { level: woodLevel },
-        );
-
-        const mainVineLeft = bezierCurveWithBricks(
-          "smallWood",
-          {
-            segments: vineSegmentsLeft,
-            spacing: 26,
-            rotationOffset: Math.PI / 2,
-          },
-          { level: woodLevel },
-        );
-
-        const mainVineLeftCenter = bezierCurveWithBricks(
-          "smallWood",
-          {
-            segments: vineSegmentsLeftCenter,
-            spacing: 26,
-            rotationOffset: Math.PI / 2,
-          },
-          { level: woodLevel },
-        );
-
-        const leafLeft = bezierPolygonWithBricks(
-          "smallOrganic",
-          {
-            outline: leafOutlineLeft,
-            spacing: 26,
-            sampleStep: 12,
-            alignToEdge: true,
-          },
-          { level: organicLevel },
-        );
-
-        const leafMid = bezierPolygonWithBricks(
-          "smallOrganic",
-          {
-            outline: leafOutlineMid,
-            spacing: 24,
-            sampleStep: 12,
-            alignToEdge: true,
-          },
-          { level: organicLevel },
-        );
-
-        const leafRight = bezierPolygonWithBricks(
-          "smallOrganic",
-          {
-            outline: leafOutlineRight,
-            spacing: 28,
-            sampleStep: 12,
-            alignToEdge: true,
-          },
-          { level: organicLevel },
-        );
-
-        return [mainVineRight, mainVineLeft, mainVineLeftCenter, 
-          // left wine
-          bezierPolygonWithBricks(
-            "smallOrganic",
-            {
-              outline: transformBezierOutline(baseLeafOutline, {
-                position: { x: center.x - 120, y: center.y + 250 },
-                scale: 0.9,
-                rotation: -2.95,
-              }),
-              spacing: 22,
-              sampleStep: 12,
-              alignToEdge: true,
-            },
-            { level: organicLevel },
-          ),
-          bezierPolygonWithBricks(
-            "smallOrganic",
-            {
-              outline: transformBezierOutline(baseLeafOutline, {
-                position: { x: center.x - 240, y: center.y + 100 },
-                scale: 0.9,
-                rotation: -3.25,
-              }),
-              spacing: 22,
-              sampleStep: 12,
-              alignToEdge: true,
-            },
-            { level: organicLevel },
-          ),
-          bezierPolygonWithBricks(
-            "smallOrganic",
-            {
-              outline: transformBezierOutline(baseLeafOutline, {
-                position: { x: center.x - 390, y: center.y + 10 },
-                scale: 0.9,
-                rotation: -3.75,
-              }),
-              spacing: 22,
-              sampleStep: 12,
-              alignToEdge: true,
-            },
-            { level: organicLevel },
-          ),
-          bezierPolygonWithBricks(
-            "smallOrganic",
-            {
-              outline: transformBezierOutline(baseLeafOutline, {
-                position: { x: center.x - 530, y: center.y - 50 },
-                scale: 0.9,
-                rotation: -3.75,
-              }),
-              spacing: 22,
-              sampleStep: 12,
-              alignToEdge: true,
-            },
-            { level: organicLevel },
-          ),
-          bezierPolygonWithBricks(
-            "smallOrganic",
-            {
-              outline: transformBezierOutline(baseLeafOutline, {
-                position: { x: center.x - 620, y: center.y - 270 },
-                scale: 0.9,
-                rotation: -2.55,
-              }),
-              spacing: 22,
-              sampleStep: 12,
-              alignToEdge: true,
-            },
-            { level: organicLevel },
-          ),
-          bezierPolygonWithBricks(
-            "smallOrganic",
-            {
-              outline: transformBezierOutline(baseLeafOutline, {
-                position: { x: center.x - 390, y: center.y - 240 },
-                scale: 0.9,
-                rotation: -1.15,
-              }),
-              spacing: 22,
-              sampleStep: 12,
-              alignToEdge: true,
-            },
-            { level: organicLevel },
-          ),
-          bezierPolygonWithBricks(
-            "smallOrganic",
-            {
-              outline: transformBezierOutline(baseLeafOutline, {
-                position: { x: center.x - 250, y: center.y - 160 },
-                scale: 0.9,
-                rotation: -1.35,
-              }),
-              spacing: 22,
-              sampleStep: 12,
-              alignToEdge: true,
-            },
-            { level: organicLevel },
-          ),
-          bezierPolygonWithBricks(
-            "smallOrganic",
-            {
-              outline: transformBezierOutline(baseLeafOutline, {
-                position: { x: center.x - 100, y: center.y - 50 },
-                scale: 0.9,
-                rotation: -1.35,
-              }),
-              spacing: 22,
-              sampleStep: 12,
-              alignToEdge: true,
-            },
-            { level: organicLevel },
-          ),
-          bezierPolygonWithBricks(
-            "smallOrganic",
-            {
-              outline: transformBezierOutline(baseLeafOutline, {
-                position: { x: center.x + 20, y: center.y + 90 },
-                scale: 0.9,
-                rotation: -1.25,
-              }),
-              spacing: 22,
-              sampleStep: 12,
-              alignToEdge: true,
-            },
-            { level: organicLevel },
-          ),
-          // medium wine
-          bezierPolygonWithBricks(
-            "smallOrganic",
-            {
-              outline: transformBezierOutline(baseLeafOutline, {
-                position: { x: center.x + 240, y: center.y + 100 },
-                scale: 0.9,
-                rotation: -0.15,
-              }),
-              spacing: 22,
-              sampleStep: 12,
-              alignToEdge: true,
-            },
-            { level: organicLevel },
-          ),
-          bezierPolygonWithBricks(
-            "smallOrganic",
-            {
-              outline: transformBezierOutline(baseLeafOutline, {
-                position: { x: center.x + 310, y: center.y - 20 },
-                scale: 0.9,
-                rotation: -0.15,
-              }),
-              spacing: 22,
-              sampleStep: 12,
-              alignToEdge: true,
-            },
-            { level: organicLevel },
-          ),
-          bezierPolygonWithBricks(
-            "smallOrganic",
-            {
-              outline: transformBezierOutline(baseLeafOutline, {
-                position: { x: center.x + 240, y: center.y - 160 },
-                scale: 0.9,
-                rotation: -0.95,
-              }),
-              spacing: 22,
-              sampleStep: 12,
-              alignToEdge: true,
-            },
-            { level: organicLevel },
-          ),
-          bezierPolygonWithBricks(
-            "smallOrganic",
-            {
-              outline: transformBezierOutline(baseLeafOutline, {
-                position: { x: center.x + 140, y: center.y - 270 },
-                scale: 0.9,
-                rotation: -0.85,
-              }),
-              spacing: 22,
-              sampleStep: 12,
-              alignToEdge: true,
-            },
-            { level: organicLevel },
-          ),
-          bezierPolygonWithBricks(
-            "smallOrganic",
-            {
-              outline: transformBezierOutline(baseLeafOutline, {
-                position: { x: center.x + 150, y: center.y - 380 },
-                scale: 0.9,
-                rotation: -0.45,
-              }),
-              spacing: 22,
-              sampleStep: 12,
-              alignToEdge: true,
-            },
-            { level: organicLevel },
-          ),
-          bezierPolygonWithBricks(
-            "smallOrganic",
-            {
-              outline: transformBezierOutline(baseLeafOutline, {
-                position: { x: center.x - 10, y: center.y - 380 },
-                scale: 0.9,
-                rotation: -2.05,
-              }),
-              spacing: 22,
-              sampleStep: 12,
-              alignToEdge: true,
-            },
-            { level: organicLevel },
-          ),
-          bezierPolygonWithBricks(
-            "smallOrganic",
-            {
-              outline: transformBezierOutline(baseLeafOutline, {
-                position: { x: center.x - 70, y: center.y - 280 },
-                scale: 0.9,
-                rotation: -2.75,
-              }),
-              spacing: 22,
-              sampleStep: 12,
-              alignToEdge: true,
-            },
-            { level: organicLevel },
-          ),
-          bezierPolygonWithBricks(
-            "smallOrganic",
-            {
-              outline: transformBezierOutline(baseLeafOutline, {
-                position: { x: center.x + 10, y: center.y - 140 },
-                scale: 0.9,
-                rotation: -3.05,
-              }),
-              spacing: 22,
-              sampleStep: 12,
-              alignToEdge: true,
-            },
-            { level: organicLevel },
-          ),
-          // left wine
-          bezierPolygonWithBricks(
-            "smallOrganic",
-            {
-              outline: transformBezierOutline(baseLeafOutline, {
-                position: { x: center.x + 110, y: center.y + 190 },
-                scale: 0.9,
-                rotation: -0.35,
-              }),
-              spacing: 22,
-              sampleStep: 12,
-              alignToEdge: true,
-            },
-            { level: organicLevel },
-          ),
-          bezierPolygonWithBricks(
-            "smallOrganic",
-            {
-              outline: transformBezierOutline(baseLeafOutline, {
-                position: { x: center.x + 510, y: center.y - 50 },
-                scale: 0.9,
-                rotation: 0.55,
-              }),
-              spacing: 22,
-              sampleStep: 12,
-              alignToEdge: true,
-            },
-            { level: organicLevel },
-          ),
-          bezierPolygonWithBricks(
-            "smallOrganic",
-            {
-              outline: transformBezierOutline(baseLeafOutline, {
-                position: { x: center.x + 620, y: center.y - 130 },
-                scale: 0.9,
-                rotation: 0.55,
-              }),
-              spacing: 22,
-              sampleStep: 12,
-              alignToEdge: true,
-            },
-            { level: organicLevel },
-          ),
-          bezierPolygonWithBricks(
-            "smallOrganic",
-            {
-              outline: transformBezierOutline(baseLeafOutline, {
-                position: { x: center.x + 640, y: center.y - 310 },
-                scale: 0.9,
-                rotation: -1.05,
-              }),
-              spacing: 22,
-              sampleStep: 12,
-              alignToEdge: true,
-            },
-            { level: organicLevel },
-          ),
-          bezierPolygonWithBricks(
-            "smallOrganic",
-            {
-              outline: transformBezierOutline(baseLeafOutline, {
-                position: { x: center.x + 520, y: center.y - 310 },
-                scale: 0.9,
-                rotation: -1.75,
-              }),
-              spacing: 22,
-              sampleStep: 12,
-              alignToEdge: true,
-            },
-            { level: organicLevel },
-          ),
-          bezierPolygonWithBricks(
-            "smallOrganic",
-            {
-              outline: transformBezierOutline(baseLeafOutline, {
-                position: { x: center.x + 410, y: center.y - 220 },
-                scale: 0.9,
-                rotation: -1.75,
-              }),
-              spacing: 22,
-              sampleStep: 12,
-              alignToEdge: true,
-            },
-            { level: organicLevel },
-          ),
-
-        ];
-      },
-      playerUnits: [
-        {
-          type: "bluePentagon",
-          position: { ...spawnPoint },
-        },
-      ],
-      unlockedBy: [
-        {
-          type: "map",
-          id: "stoneCottage",
-          level: 1,
-        },
-      ],
-      mapsRequired: { stoneCottage: 1 },
-      maxLevel: 2,
-    } satisfies MapConfig;
-  })(),
-  mine: (() => {
-    const size: SceneSize = { width: 1200, height: 1300 };
-    const center: SceneVector2 = { x: size.width / 2, y: (size.height-100) / 2 };
-    const shaftRadius = 320;
-    const wallThickness = 120;
-    const entryWidth = 140;
-    const spawnPoint: SceneVector2 = { x: center.x, y: size.height - 260 };
-
-    const createSupport = (
-      angle: number,
-      length: number,
-      width: number,
-    ): SceneVector2[] => {
-      const dx = Math.cos(angle) * length;
-      const dy = Math.sin(angle) * length;
-      const px = -Math.sin(angle) * width;
-      const py = Math.cos(angle) * width;
-      return [
-        { x: center.x - dx + px, y: center.y - dy + py },
-        { x: center.x - dx - px, y: center.y - dy - py },
-        { x: center.x + dx - px, y: center.y + dy - py },
-        { x: center.x + dx + px, y: center.y + dy + py },
-      ];
-    };
-
-    return {
-      name: "Collapsed Mine",
-      size,
-      spawnPoints: [spawnPoint],
-      nodePosition: { x: 4, y: 5 },
-      icon: "mine.png",
-      bricks: ({ mapLevel }) => {
-        const baseLevel = Math.max(0, Math.floor(mapLevel));
-        const wallLevel = baseLevel + 1;
-
-        const ironWalls = circleWithBricks(
-          "smallIron",
-          {
-            center,
-            innerRadius: shaftRadius,
-            outerRadius: shaftRadius + wallThickness,
-          },
-          { level: wallLevel },
-        );
-
-        const coalVein = circleWithBricks(
-          "smallCoal",
-          {
-            center,
-            innerRadius: 0,
-            outerRadius: shaftRadius - 40,
-          },
-          { level: baseLevel },
-        );
-
-        const entryTunnel = polygonWithBricks(
-          "smallSquareGray",
-          {
-            vertices: [
-              { x: center.x - entryWidth / 2, y: spawnPoint.y },
-              { x: center.x + entryWidth / 2, y: spawnPoint.y },
-              { x: center.x + entryWidth / 2, y: center.y + shaftRadius - 20 },
-              { x: center.x - entryWidth / 2, y: center.y + shaftRadius - 20 },
-            ],
-          },
-          { level: wallLevel },
-        );
-
-        const supports = [0, Math.PI / 2, Math.PI, (3 * Math.PI) / 2].map(
-          (angle) =>
-            polygonWithBricks(
-              "smallIron",
-              {
-                vertices: createSupport(
-                  angle,
-                  shaftRadius + wallThickness * 0.45,
-                  30,
-                ),
-              },
-              { level: wallLevel },
-            ),
-        );
-
-        return [ironWalls, coalVein, entryTunnel, ...supports];
-      },
-      enemies: ({ mapLevel }) => {
-        const baseLevel = Math.max(0, Math.floor(mapLevel));
-        return [
-          {
-            type: "laserTurretEnemy",
-            level: baseLevel,
-            position: { x: 100, y: center.y },
-          } satisfies EnemySpawnData,
-          {
-            type: "laserTurretEnemy",
-            level: baseLevel,
-            position: { x: size.width - 100, y: center.y },
-          } satisfies EnemySpawnData,
-        ];
-      },
-      playerUnits: [
-        {
-          type: "bluePentagon",
-          position: { ...spawnPoint },
-        },
-      ],
-      unlockedBy: [
-        {
-          type: "map",
-          id: "spruce",
-          level: 1,
-        },
-      ],
-      mapsRequired: { spruce: 1 },
-      maxLevel: 1,
-    } satisfies MapConfig;
-  })(),
-  uranium_fields: (() => {
-    const size: SceneSize = { width: 1200, height: 1200 };
-    const center: SceneVector2 = { x: size.width / 2, y: size.height / 2 };
-    const spawnPoint: SceneVector2 = { x: 100, y: 100 };
-
-    return {
-      name: "Uranium Fields",
-      size,
-      spawnPoints: [spawnPoint],
-      nodePosition: { x: 4, y: 6 },
-      icon: "uranium_fields.png",
-      lockedForDemo: true,
-      mapEffects: ["radioactivity"],
-      bricks: ({ mapLevel }) => {
-
-        const numPetals = 3;
-        const slicedTrianles = [];
-
-        const radius = 300
-
-        for (let i = 0; i < numPetals; i++) {
-          const angleWidth = Math.PI / (numPetals + 1);
-          const angle = (i / numPetals) * 2 * Math.PI - Math.PI / 2;
-          const leftAngle = angle - angleWidth / 2;
-          const rightAngle = angle + angleWidth / 2;
-          const xLeft = center.x + Math.cos(leftAngle) * radius;
-          const yLeft = center.y + Math.sin(leftAngle) * radius;
-          const xRight = center.x + Math.cos(rightAngle) * radius;
-          const yRight = center.y + Math.sin(rightAngle) * radius;
-          slicedTrianles.push({ 
-            vertices: [
-              { x: center.x, y: center.y }, 
-              { x: xLeft, y: yLeft }, 
-              { x: xRight, y: yRight }
-            ] 
-          });
-        }
-
-        const baseLevel = Math.max(0, Math.floor(mapLevel));
-        return [
-          ...slicedTrianles.map((triangle) =>
-            polygonWithBricks(
-              "uraniumBrick",
-              triangle,
-              { level: baseLevel },
-            ),
-          ),
-          circleWithBricks(
-            "uraniumBrick",
-            {
-              center,
-              innerRadius: radius + 50,
-              outerRadius: radius + 120,
-            },
-            { level: baseLevel },
-          ),
-        ];
-      },
-      playerUnits: [
-        {
-          type: "bluePentagon",
-          position: { ...spawnPoint },
-        },
-      ],
-      unlockedBy: [
-        {
-          type: "map",
-          id: "mine",
-          level: 1,
-        },
-      ],
-      mapsRequired: { mine: 1 },
-      maxLevel: 1,
-    } satisfies MapConfig;
-  })(),
-  adit: (() => {
-    const size: SceneSize = { width: 1500, height: 1200 };
-    const centerX = size.width / 2;
-    const centerY = size.height / 2;
-    // Spawn point in center of central room (400x400)
-    const spawnPoint: SceneVector2 = { x: centerX, y: centerY };
-
-    const createRectangle = (
-      x: number,
-      y: number,
-      width: number,
-      height: number,
-    ): SceneVector2[] => [
-      { x, y },
-      { x: x + width, y },
-      { x: x + width, y: y + height },
-      { x, y: y + height },
-    ];
-
-    return {
-      name: "Adit Corridors",
-      size,
-      spawnPoints: [spawnPoint],
-      nodePosition: { x: 4, y: 3 },
-      icon: "adit.png",
-      bricks: ({ mapLevel }) => {
-        const baseLevel = Math.max(0, Math.floor(mapLevel));
-        const ironLevel = baseLevel;
-        const wallThickness = 40;
-        const corridorWidth = 250; // 200-300px as requested
-
-        // === CENTRAL ROOM (400x400) ===
-        const centralRoomSize = 400;
-        const centralRoomX = centerX - centralRoomSize / 2;
-        const centralRoomY = centerY - centralRoomSize / 2;
-
-        // === MAZE WALLS (iron) ===
-        const walls: ReturnType<typeof polygonWithBricks>[] = [];
-
-        // Outer border walls
-        walls.push(
-          // Top wall
-          polygonWithBricks(
-            "compactIron",
-            {
-              vertices: createRectangle(
-                100,
-                100,
-                size.width - 200,
-                wallThickness,
-              ),
-            },
-            { level: ironLevel },
-          ),
-          // Bottom wall
-          polygonWithBricks(
-            "compactIron",
-            {
-              vertices: createRectangle(
-                100,
-                size.height - 100 - wallThickness,
-                size.width - 200,
-                wallThickness,
-              ),
-            },
-            { level: ironLevel },
-          ),
-          // Left wall
-          polygonWithBricks(
-            "compactIron",
-            {
-              vertices: createRectangle(
-                100,
-                100,
-                wallThickness,
-                size.height - 200,
-              ),
-            },
-            { level: ironLevel },
-          ),
-          // Right wall
-          polygonWithBricks(
-            "compactIron",
-            {
-              vertices: createRectangle(
-                size.width - 100 - wallThickness,
-                100,
-                wallThickness,
-                size.height - 200,
-              ),
-            },
-            { level: ironLevel },
-          ),
-        );
-
-        // Central room walls (surrounding the 400x400 room)
-        walls.push(
-          // Top wall of central room
-          polygonWithBricks(
-            "compactIron",
-            {
-              vertices: createRectangle(
-                centralRoomX - wallThickness,
-                centralRoomY - wallThickness,
-                centralRoomSize + wallThickness * 2,
-                wallThickness,
-              ),
-            },
-            { level: ironLevel },
-          ),
-          // Bottom wall of central room
-          polygonWithBricks(
-            "compactIron",
-            {
-              vertices: createRectangle(
-                centralRoomX - wallThickness,
-                centralRoomY + centralRoomSize,
-                centralRoomSize + wallThickness * 2,
-                wallThickness,
-              ),
-            },
-            { level: ironLevel },
-          ),
-          // Left wall of central room
-          polygonWithBricks(
-            "compactIron",
-            {
-              vertices: createRectangle(
-                centralRoomX - wallThickness,
-                centralRoomY,
-                wallThickness,
-                centralRoomSize,
-              ),
-            },
-            { level: ironLevel },
-          ),
-          // Right wall of central room
-          polygonWithBricks(
-            "compactIron",
-            {
-              vertices: createRectangle(
-                centralRoomX + centralRoomSize,
-                centralRoomY,
-                wallThickness,
-                centralRoomSize,
-              ),
-            },
-            { level: ironLevel },
-          ),
-        );
-
-        // Maze corridors - create walls around corridors (corridors are open spaces)
-        // Top corridor walls (vertical walls on sides of corridor)
-        const topCorridorY = 100 + wallThickness;
-        const topCorridorHeight = centralRoomY - topCorridorY;
-        walls.push(
-          // Left wall of top corridor
-          polygonWithBricks(
-            "compactIron",
-            {
-              vertices: createRectangle(
-                centerX - corridorWidth / 2 - wallThickness,
-                topCorridorY,
-                wallThickness,
-                topCorridorHeight,
-              ),
-            },
-            { level: ironLevel },
-          ),
-          // Right wall of top corridor
-          polygonWithBricks(
-            "compactIron",
-            {
-              vertices: createRectangle(
-                centerX + corridorWidth / 2,
-                topCorridorY,
-                wallThickness,
-                topCorridorHeight,
-              ),
-            },
-            { level: ironLevel },
-          ),
-        );
-
-        // Bottom corridor walls
-        const bottomCorridorY = centralRoomY + centralRoomSize + wallThickness;
-        const bottomCorridorHeight =
-          size.height - 100 - wallThickness - bottomCorridorY;
-        walls.push(
-          // Left wall of bottom corridor
-          polygonWithBricks(
-            "compactIron",
-            {
-              vertices: createRectangle(
-                centerX - corridorWidth / 2 - wallThickness,
-                bottomCorridorY,
-                wallThickness,
-                bottomCorridorHeight,
-              ),
-            },
-            { level: ironLevel },
-          ),
-          // Right wall of bottom corridor
-          polygonWithBricks(
-            "compactIron",
-            {
-              vertices: createRectangle(
-                centerX + corridorWidth / 2,
-                bottomCorridorY,
-                wallThickness,
-                bottomCorridorHeight,
-              ),
-            },
-            { level: ironLevel },
-          ),
-        );
-
-        // Left corridor walls (horizontal walls on top/bottom of corridor)
-        const leftCorridorX = 100 + wallThickness;
-        const leftCorridorWidth = centralRoomX - leftCorridorX;
-        walls.push(
-          // Top wall of left corridor
-          polygonWithBricks(
-            "compactIron",
-            {
-              vertices: createRectangle(
-                leftCorridorX,
-                centerY - corridorWidth / 2 - wallThickness,
-                leftCorridorWidth,
-                wallThickness,
-              ),
-            },
-            { level: ironLevel },
-          ),
-          // Bottom wall of left corridor
-          polygonWithBricks(
-            "compactIron",
-            {
-              vertices: createRectangle(
-                leftCorridorX,
-                centerY + corridorWidth / 2,
-                leftCorridorWidth,
-                wallThickness,
-              ),
-            },
-            { level: ironLevel },
-          ),
-        );
-
-        // Right corridor walls
-        const rightCorridorX = centralRoomX + centralRoomSize + wallThickness;
-        const rightCorridorWidth =
-          size.width - 100 - wallThickness - rightCorridorX;
-        walls.push(
-          // Top wall of right corridor
-          polygonWithBricks(
-            "compactIron",
-            {
-              vertices: createRectangle(
-                rightCorridorX,
-                centerY - corridorWidth / 2 - wallThickness,
-                rightCorridorWidth,
-                wallThickness,
-              ),
-            },
-            { level: ironLevel },
-          ),
-          // Bottom wall of right corridor
-          polygonWithBricks(
-            "compactIron",
-            {
-              vertices: createRectangle(
-                rightCorridorX,
-                centerY + corridorWidth / 2,
-                rightCorridorWidth,
-                wallThickness,
-              ),
-            },
-            { level: ironLevel },
-          ),
-        );
-
-        // Additional maze walls (creating dead ends and paths)
-        // Horizontal dividers in corners
-        walls.push(
-          // Top-left corner divider
-          polygonWithBricks(
-            "compactIron",
-            {
-              vertices: createRectangle(200, 300, 300, wallThickness),
-            },
-            { level: ironLevel },
-          ),
-          // Top-right corner divider
-          polygonWithBricks(
-            "compactIron",
-            {
-              vertices: createRectangle(1000, 300, 300, wallThickness),
-            },
-            { level: ironLevel },
-          ),
-          // Bottom-left corner divider
-          polygonWithBricks(
-            "compactIron",
-            {
-              vertices: createRectangle(
-                200,
-                size.height - 100 - wallThickness - 300,
-                300,
-                wallThickness,
-              ),
-            },
-            { level: ironLevel },
-          ),
-          // Bottom-right corner divider
-          polygonWithBricks(
-            "compactIron",
-            {
-              vertices: createRectangle(
-                1000,
-                size.height - 100 - wallThickness - 300,
-                300,
-                wallThickness,
-              ),
-            },
-            { level: ironLevel },
-          ),
-        );
-
-        // Vertical dividers in corners
-        walls.push(
-          // Top-left vertical
-          polygonWithBricks(
-            "compactIron",
-            {
-              vertices: createRectangle(
-                400,
-                100 + wallThickness,
-                wallThickness,
-                200,
-              ),
-            },
-            { level: ironLevel },
-          ),
-          // Top-right vertical
-          polygonWithBricks(
-            "compactIron",
-            {
-              vertices: createRectangle(
-                1100,
-                100 + wallThickness,
-                wallThickness,
-                200,
-              ),
-            },
-            { level: ironLevel },
-          ),
-          // Bottom-left vertical
-          polygonWithBricks(
-            "compactIron",
-            {
-              vertices: createRectangle(
-                400,
-                size.height - 100 - wallThickness - 200,
-                wallThickness,
-                200,
-              ),
-            },
-            { level: ironLevel },
-          ),
-          // Bottom-right vertical
-          polygonWithBricks(
-            "compactIron",
-            {
-              vertices: createRectangle(
-                1100,
-                size.height - 100 - wallThickness - 200,
-                wallThickness,
-                200,
-              ),
-            },
-            { level: ironLevel },
-          ),
-        );
-
-        return walls;
-      },
-      playerUnits: [
-        {
-          type: "bluePentagon",
-          position: { ...spawnPoint },
-        },
-      ],
-      unlockedBy: [
-        {
-          type: "map",
-          id: "oldForge",
-          level: 1,
-        },
-      ],
-      mapsRequired: { oldForge: 1 },
-      maxLevel: 2,
-    } satisfies MapConfig;
-  })(),
-  wire: (() => {
-    const size: SceneSize = { width: 1500, height: 1500 };
-    const cableCenters: readonly SceneVector2[] = [
-      { x: 350, y: 350 },
-      { x: 520, y: 260 },
-      { x: 780, y: 420 },
-      { x: 1030, y: 360 },
-      { x: 1280, y: 520 },
-      { x: 1120, y: 820 },
-      { x: 860, y: 960 },
-      { x: 620, y: 900 },
-      { x: 420, y: 1080 },
-    ];
-
-    const spawnPoint: SceneVector2 = { x: 180, y: 300 };
-
-    return {
-      name: "Wire",
-      size,
-      spawnPoints: [spawnPoint],
-      nodePosition: { x: 4, y: 1 },
-      icon: "wire.png",
-      bricks: ({ mapLevel }) => {
-        const baseLevel = Math.max(0, Math.floor(mapLevel));
-        const outerLevel = baseLevel + 2;
-        const outerRadius = 90;
-        const innerRadius = 40;
-
-        const outerSegments = cableCenters.flatMap((center, index) => {
-          const circle = circleWithBricks(
-            "smallSquareGray",
-            {
-              center,
-              innerRadius: 0,
-              outerRadius,
-            },
-            { level: outerLevel },
-          );
-
-          if (index >= cableCenters.length - 1) {
-            return [circle];
-          }
-
-          const nextCenter = cableCenters[index + 1];
-          if (!nextCenter) {
-            return [circle];
-          }
-
-          const connector = connectorWithBricks(
-            "smallSquareGray",
-            {
-              start: center,
-              end: nextCenter,
-              width: outerRadius * 2,
-            },
-            { level: outerLevel },
-          );
-
-          return [circle, connector];
-        });
-
-        const innerSegments = cableCenters.flatMap((center, index) => {
-          const circle = circleWithBricks(
-            "smallCopper",
-            {
-              center,
-              innerRadius: 0,
-              outerRadius: innerRadius,
-            },
-            { level: baseLevel },
-          );
-
-          if (index >= cableCenters.length - 1) {
-            return [circle];
-          }
-
-          const nextCenter = cableCenters[index + 1];
-          if (!nextCenter) {
-            return [circle];
-          }
-
-          const connector = connectorWithBricks(
-            "smallCopper",
-            {
-              start: center,
-              end: nextCenter,
-              width: innerRadius * 2,
-            },
-            { level: baseLevel },
-          );
-
-          return [circle, connector];
-        });
-
-        return [...outerSegments, ...innerSegments];
-      },
-      playerUnits: [
-        {
-          type: "bluePentagon",
-          position: { ...spawnPoint },
-        },
-      ],
-      unlockedBy: [
-        {
-          type: "map",
-          id: "oldForge",
-          level: 1,
-        },
-      ],
-      mapsRequired: { oldForge: 1 },
-      maxLevel: 1,
-    } satisfies MapConfig;
-  })(),
-  coil: (() => {
-    const size: SceneSize = { width: 1500, height: 1500 };
-    const center: SceneVector2 = { x: size.width / 2, y: size.height / 2 };
-    const spawnPoint: SceneVector2 = { x: center.x - 650, y: center.y };
-    const outerRadius = 520;
-    const innerRadius = 390;
-    const wireWidth = outerRadius - innerRadius + 20;
-
-    return {
-      name: "Coil",
-      size,
-      icon: "coil.png",
-      spawnPoints: [spawnPoint],
-      nodePosition: { x: 5, y: 1 },
-      lockedForDemo: true,
-      bricks: ({ mapLevel }) => {
-        const baseLevel = Math.max(0, Math.floor(mapLevel));
-        const ringLevel = baseLevel + 3;
-        const wireLevel = baseLevel + 2;
-
-        const base = circleWithBricks(
-          "smallSquareGray",
-          {
-            center,
-            innerRadius,
-            outerRadius,
-          },
-          { level: ringLevel },
-        );
-
-        const wires = [];
-
-        for (let i = 0; i < 60; i++) {
-          const angle = (i / 60) * Math.PI * 2;
-          const position = {
-            x: center.x + outerRadius + 50 * Math.cos(angle),
-            y: center.y + outerRadius + 50 * Math.sin(angle),
-          };
-          const copperWire = bezierCurveWithBricks(
-            "smallCopper",
-            {
-              segments: [
-                {
-                  start: { x: center.x + innerRadius * Math.cos(angle), y: center.y + innerRadius * Math.sin(angle) },
-                  control1: { x: center.x + innerRadius * Math.cos(angle+0.1) + (wireWidth/3) * Math.cos(angle+0.1), y: center.y + innerRadius * Math.sin(angle+0.1) + (wireWidth/3) * Math.sin(angle+0.1) },
-                  control2: { x: center.x + innerRadius * Math.cos(angle+0.1) + (2*wireWidth/3) * Math.cos(angle+0.1), y: center.y + innerRadius * Math.sin(angle+0.1) + (2*wireWidth/3) * Math.sin(angle+0.1) },
-                  end: { x: center.x + (innerRadius +wireWidth) * Math.cos(angle), y: center.y + (innerRadius +wireWidth) * Math.sin(angle) },
-                }
-              ],
-              spacing: 26,
-              rotationOffset: Math.PI / 2,
-            },
-            { level: wireLevel },
-          );
-          wires.push(copperWire);
-        }
-        return [base, ...wires];
-      },
-      /* enemies: ({ mapLevel }) => {
-        const level = Math.max(1, Math.floor(mapLevel));
-        const offsets: SceneVector2[] = [
-          { x: -100, y: -100 },
-          { x: 100, y: -100 },
-          { x: 100, y: 100 },
-          { x: -100, y: 100 },
-        ];
-        return offsets.map((offset) => ({
-          type: "burstTurretEnemy",
-          level,
-          position: {
-            x: center.x + offset.x,
-            y: center.y + offset.y,
-          },
-        }));
-      },*/
-      playerUnits: [
-        {
-          type: "bluePentagon",
-          position: { ...spawnPoint },
-        },
-      ],
-      unlockedBy: [
-        {
-          type: "map",
-          id: "wire",
-          level: 1,
-        },
-      ],
-      mapsRequired: { wire: 1 },
-      maxLevel: 1,
-    } satisfies MapConfig;
-  })(),
-  silverRing: (() => {
-    const size: SceneSize = { width: 1500, height: 1500 };
-    const center: SceneVector2 = { x: size.width / 2, y: size.height / 2 };
-    const spawnPoint: SceneVector2 = { x: center.x - 650, y: center.y };
-    const outerRadius = 520;
-    const innerRadius = 360;
-    const gemRadius = 60;
-
-    return {
-      name: "Silver Ring",
-      size,
-      icon: "silver_ring.png",
-      spawnPoints: [spawnPoint],
-      nodePosition: { x: 4, y: 0 },
-      bricks: ({ mapLevel }) => {
-        const baseLevel = Math.max(0, Math.floor(mapLevel));
-        const ringLevel = baseLevel;
-        const gemLevel = baseLevel + 1;
-
-        const silverRing = circleWithBricks(
-          "smallSilver",
-          {
-            center,
-            innerRadius,
-            outerRadius,
-          },
-          { level: ringLevel },
-        );
-
-        const copperGem = circleWithBricks(
-          "smallCopper",
-          {
-            center: { x: center.x + outerRadius + 50, y: center.y },
-            innerRadius: 0,
-            outerRadius: gemRadius * 1.25,
-          },
-          { level: gemLevel },
-        );
-        const copperGem2 = circleWithBricks(
-          "smallCopper",
-          {
-            center: { x: center.x + outerRadius - 25, y: center.y + 50 },
-            innerRadius: 0,
-            outerRadius: gemRadius,
-          },
-          { level: gemLevel },
-        );
-        const copperGem3 = circleWithBricks(
-          "smallCopper",
-          {
-            center: { x: center.x + outerRadius - 25, y: center.y - 50 },
-            innerRadius: 0,
-            outerRadius: gemRadius,
-          },
-          { level: gemLevel },
-        );
-
-        return [silverRing, copperGem, copperGem2, copperGem3];
-      },
-      /* enemies: ({ mapLevel }) => {
-        const level = Math.max(1, Math.floor(mapLevel));
-        const offsets: SceneVector2[] = [
-          { x: -100, y: -100 },
-          { x: 100, y: -100 },
-          { x: 100, y: 100 },
-          { x: -100, y: 100 },
-        ];
-        return offsets.map((offset) => ({
-          type: "burstTurretEnemy",
-          level,
-          position: {
-            x: center.x + offset.x,
-            y: center.y + offset.y,
-          },
-        }));
-      },*/
-      playerUnits: [
-        {
-          type: "bluePentagon",
-          position: { ...spawnPoint },
-        },
-      ],
-      unlockedBy: [
-        {
-          type: "map",
-          id: "wire",
-          level: 1,
-        },
-      ],
-      mapsRequired: { wire: 1 },
-      maxLevel: 1,
-    } satisfies MapConfig;
-  })(),
-  frozenForest: (() => {
-    const size: SceneSize = { width: 1500, height: 1500 };
-    const center: SceneVector2 = { x: size.width / 2, y: size.height / 2 };
-    const spawnPoint: SceneVector2 = { x: 200, y: 200 };
-    const lakeRadius = 450;
-
-    const createRectangle = (
-      x: number,
-      y: number,
-      width: number,
-      height: number,
-    ): SceneVector2[] => [
-      { x, y },
-      { x: x + width, y },
-      { x: x + width, y: y + height },
-      { x, y: y + height },
-    ];
-
-    const createTriangle = (
-      baseCenter: SceneVector2,
-      width: number,
-      height: number,
-    ): SceneVector2[] => [
-      { x: baseCenter.x, y: baseCenter.y - height },
-      { x: baseCenter.x + width / 2, y: baseCenter.y },
-      { x: baseCenter.x - width / 2, y: baseCenter.y },
-    ];
-
-    const treeConfigs: readonly { base: SceneVector2; scale: number }[] = [
-      { base: { x: 300, y: 300 }, scale: 0.8 },
-      { base: { x: 1200, y: 400 }, scale: 0.9 },
-      { base: { x: 200, y: 900 }, scale: 0.85 },
-      { base: { x: 1100, y: 1100 }, scale: 0.95 },
-      { base: { x: 1300, y: 800 }, scale: 0.75 },
-      { base: { x: 400, y: 1200 }, scale: 0.8 },
-    ];
-
-    return {
-      name: "Frozen Forest",
-      size,
-      spawnPoints: [spawnPoint],
-      nodePosition: { x: 3, y: -1 },
-      icon: "frozen_forest.png",
-      lockedForDemo: true,
-      bricks: ({ mapLevel }) => {
-        const baseLevel = Math.max(0, Math.floor(mapLevel));
-        const iceLevel = baseLevel;
-        const treeTrunkLevel = baseLevel;
-        const treeCanopyLevel = baseLevel + 1;
-
-        const frozenLake = circleWithBricks(
-          "smallIce",
-          {
-            center,
-            innerRadius: 0,
-            outerRadius: lakeRadius,
-          },
-          { level: iceLevel },
-        );
-
-        const bush1 = brickTreeDeterministic({
-          origin: { x: 1180, y: 480 },
-          size: 260,
-          mainBranchThickness: 19,
-          twigThickness: 18,
-          trunkThickness: 38,
-          sideBranchesCount: 4,
-          branchingsPerSide: 2,
-          topBranchesCount: 2,
-          minSideAngleDeg: 60,
-          sideSpreadRad: Math.PI * 0.1,
-          topSpreadRad: Math.PI * 0.22,
-          twigSpreadRad: Math.PI * 0.25,
-          branchStartT: 0.25,
-          bendTrunkPx: 22,
-          bendTopPx: 10,
-          bendSidePx: 18,
-          bendTwigPx: 10,
-          brickType: "smallWood",
-          brickLevel: treeCanopyLevel,
-          spacing: 26,
-          angleJitterRad: 0.72,
-          lenJitter: 0.12,
-          bendJitter: 0.30,
-          verticalBias: 0.22,
-          //heightJitterT: 0.27,
-        });
-
-        const bush2 = brickTreeDeterministic({
-          origin: { x: 1250, y: 980 },
-          size: 260,
-          mainBranchThickness: 19,
-          twigThickness: 18,
-          trunkThickness: 38,
-          sideBranchesCount: 3,
-          branchingsPerSide: 2,
-          topBranchesCount: 3,
-          minSideAngleDeg: 60,
-          sideSpreadRad: Math.PI * 0.4,
-          topSpreadRad: Math.PI * 0.42,
-          twigSpreadRad: Math.PI * 0.15,
-          branchStartT: 0.25,
-          bendTrunkPx: 22,
-          bendTopPx: 10,
-          bendSidePx: 18,
-          bendTwigPx: 10,
-          brickType: "smallWood",
-          brickLevel: treeCanopyLevel,
-          spacing: 26,
-          angleJitterRad: 0.72,
-          lenJitter: 0.12,
-          bendJitter: 0.30,
-          verticalBias: 0.22,
-          //heightJitterT: 0.27,
-        });
-
-        const bush3 = brickTreeDeterministic({
-          origin: { x: 1150, y: 1180 },
-          size: 160,
-          mainBranchThickness: 19,
-          twigThickness: 18,
-          trunkThickness: 38,
-          sideBranchesCount: 2,
-          branchingsPerSide: 3,
-          topBranchesCount: 3,
-          minSideAngleDeg: 60,
-          branchesLenMul: 1.4,
-          sideSpreadRad: Math.PI * 0.4,
-          topSpreadRad: Math.PI * 0.42,
-          twigSpreadRad: Math.PI * 0.15,
-          branchStartT: 0.05,
-          branchEndT: 0.25,
-          bendTrunkPx: 22,
-          bendTopPx: 10,
-          bendSidePx: 18,
-          bendTwigPx: 10,
-          brickType: "smallWood",
-          brickLevel: treeCanopyLevel,
-          spacing: 26,
-          angleJitterRad: 0.72,
-          lenJitter: 0.12,
-          bendJitter: 0.30,
-          verticalBias: 0.22,
-          //heightJitterT: 0.27,
-        });
-
-        const bush4 = brickTreeDeterministic({
-          origin: { x: 250, y: 980 },
-          size: 360,
-          mainBranchThickness: 19,
-          twigThickness: 18,
-          trunkThickness: 38,
-          sideBranchesCount: 6,
-          branchingsPerSide: 2,
-          topBranchesCount: 3,
-          minSideAngleDeg: 60,
-          sideSpreadRad: Math.PI * 0.4,
-          topSpreadRad: Math.PI * 0.42,
-          twigSpreadRad: Math.PI * 0.15,
-          branchStartT: 0.15,
-          bendTrunkPx: 22,
-          bendTopPx: 10,
-          bendSidePx: 18,
-          bendTwigPx: 10,
-          brickType: "smallWood",
-          brickLevel: treeCanopyLevel,
-          spacing: 26,
-          angleJitterRad: 0.72,
-          lenJitter: 0.12,
-          bendJitter: 0.30,
-          verticalBias: 0.22,
-          twigAt: [0.25, 0.45, 0.65],
-          //heightJitterT: 0.27,
-        });
-
-
-        return [frozenLake, ...bush1, ...bush2, ...bush3, ...bush4];
-      },
-      playerUnits: [
-        {
-          type: "bluePentagon",
-          position: { ...spawnPoint },
-        },
-      ],
-      unlockedBy: [
-        {
-          type: "map",
-          id: "silverRing",
-          level: 1,
-        },
-      ],
-      mapsRequired: { silverRing: 1 },
-      maxLevel: 1,
-    } satisfies MapConfig;
-  })(),
-  volcano: (() => {
-    const size: SceneSize = { width: 1500, height: 1500 };
-    const center: SceneVector2 = { x: size.width / 2, y: size.height / 2 };
-    const spawnPoint: SceneVector2 = { x: 200, y: size.height - 200 };
-    const volcanoBaseRadius = 400;
-    const volcanoInnerRadius = 200;
-    const magmaFlowRadius = 80;
-
-    const createRectangle = (
-      x: number,
-      y: number,
-      width: number,
-      height: number,
-    ): SceneVector2[] => [
-      { x, y },
-      { x: x + width, y },
-      { x: x + width, y: y + height },
-      { x, y: y + height },
-    ];
-
-    const magmaFlowPaths: readonly {
-      center: SceneVector2;
-      length: number;
-      angle: number;
-    }[] = [
-      {
-        center: { x: center.x - 300, y: center.y + 200 },
-        length: 200,
-        angle: 0.5,
-      },
-      {
-        center: { x: center.x + 250, y: center.y + 150 },
-        length: 180,
-        angle: -0.3,
-      },
-      { center: { x: center.x, y: center.y + 300 }, length: 220, angle: 0 },
-      {
-        center: { x: center.x - 200, y: center.y - 100 },
-        length: 150,
-        angle: 1.2,
-      },
-      {
-        center: { x: center.x + 300, y: center.y - 150 },
-        length: 170,
-        angle: -1.0,
-      },
-    ];
-
-    return {
-      name: "Volcano",
-      size,
-      spawnPoints: [spawnPoint],
-      nodePosition: { x: 5, y: 5 },
-      icon: "volcano.png",
-      lockedForDemo: true,
-      bricks: ({ mapLevel }) => {
-        const baseLevel = Math.max(0, Math.floor(mapLevel));
-        const copperLevel = baseLevel + 2;
-        const stoneLevel = baseLevel + 4;
-        const magmaLevel = baseLevel;
-
-        const volcanoBase = circleWithBricks(
-          "smallCopper",
-          {
-            center,
-            innerRadius: volcanoInnerRadius,
-            outerRadius: volcanoBaseRadius,
-          },
-          { level: copperLevel },
-        );
-
-        const volcanoCore = circleWithBricks(
-          "smallSquareGray",
-          {
-            center,
-            innerRadius: 0,
-            outerRadius: volcanoInnerRadius - 40,
-          },
-          { level: stoneLevel },
-        );
-
-        const magmaFlows = magmaFlowPaths.map((flow) =>
-          circleWithBricks(
-            "smallMagma",
-            {
-              center: flow.center,
-              innerRadius: 0,
-              outerRadius: magmaFlowRadius,
-            },
-            { level: magmaLevel },
-          ),
-        );
-
-        return [volcanoBase, volcanoCore, ...magmaFlows];
-      },
-      playerUnits: [
-        {
-          type: "bluePentagon",
-          position: { ...spawnPoint },
-        },
-      ],
-      unlockedBy: [
-        {
-          type: "map",
-          id: "mine",
-          level: 1,
-        },
-      ],
-      mapsRequired: { mine: 1 },
-      maxLevel: 1,
-    } satisfies MapConfig;
-  })(),
+  tutorialZone,
+  trainingGrounds,
+  foundations,
+  initial,
+  turretRings,
+  thicket,
+  oldForge,
+  spruce,
+  deadOak,
+  sphinx,
+  stoneCottage,
+  bezierGrove,
+  wire,
+  coil,
+  mine,
+  adit,
+  silverRing,
+  frozenForest,
+  volcano,
+  megaBrick,
+  ancientPyramids,
+  deathfulGuns,
+  deadlyTunnels,
+  encagedBeast,
+  uranium_fields: uraniumFields,
 };
 
 export const MAP_IDS = Object.keys(MAPS_DB) as MapId[];
